% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion.R
\name{accordion_mutate}
\alias{accordion_mutate}
\title{Update an accordion}
\usage{
accordion_mutate(
  id,
  target,
  ...,
  title = NULL,
  value = NULL,
  icon = NULL,
  immediate = FALSE,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{an input id.}

\item{target}{The \code{value} of an existing \code{\link[=accordion_item]{accordion_item()}}}
}
\description{
Update an accordion
}
