% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{as.tags.bslib_sidebar}
\alias{as.tags.bslib_sidebar}
\title{Render a sidebar as HTML tags}
\usage{
\method{as.tags}{bslib_sidebar}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=sidebar]{sidebar()}} object.}

\item{...}{Additional arguments passed to \code{\link[htmltools:as.tags]{htmltools::as.tags()}}.}
}
\description{
Renders the sidebar element and collapse toggle elements for a \code{\link[=sidebar]{sidebar()}} in
a \code{\link[=layout_sidebar]{layout_sidebar()}} context.
}
\keyword{internal}
