% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_sass.R
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap_sass}
\title{Compile Bootstrap 4 (or 3) SASS with (optional) theming}
\usage{
bootstrap(
  theme = bs_theme_get(),
  jquery = jquerylib::jquery_core(3),
  options = sass::sass_options(),
  minified = TRUE
)

bootstrap_sass(
  rules = list(),
  theme = bs_theme_get(),
  options = sass::sass_options()
)
}
\arguments{
\item{theme}{one of the following:
\enumerate{
\item The result of \code{\link[=bs_theme_get]{bs_theme_get()}} (recommended default).
\item A string containing a bootswatch theme and/or a Bootstrap major version. To specify
both, use the syntax \code{"theme@version"}, (e.g., \code{"cosmo@4"} for Bootstrap 4 cosmo
theme with BS3 compatibility). If no version is specified, the latest available
version is used (for more info, see \code{version} in \code{\link[=bs_theme_set]{bs_theme_set()}}).
\item \code{NULL}, which means use the latest version of Bootstrap with no custom theming.
}}

\item{jquery}{See \code{\link[jquerylib:jquery_core]{jquerylib::jquery_core()}}.}

\item{options}{Compiler options for Sass. Please specify options using
\code{\link[sass]{sass_options}}.}

\item{minified}{whether the resulting HTML dependency should minify the JS/CSS files.}
}
\description{
Use \code{bootstrap()} to obtain all of Bootstrap and it's dependencies as a list
of \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} objects.
Use \code{bootstrap_sass()} if you can assume Bootstrap already exists in your
project, but you want to leverage Bootstrap utilities (e.g., variables, functions,
or mixins) to generate additional CSS rules (as a string that can be included
as a \verb{<style>} tag via \code{tags$style(css)}).
}
\examples{
library(htmltools)

button <- tags$a(class = "btn btn-primary", href = "#", role = "button", "Hello")

preview_button <- function(x) {
  browsable(tags$body(x, button))
}

# Latest bootstrap
preview_button(bootstrap())

# Bootstrap 3
preview_button(bootstrap("3"))

# Bootswatch minty theme
preview_button(bootstrap("minty"))

# Bootswatch sketchy theme
preview_button(bootstrap("sketchy"))

# Bootswatch solar theme with BS3 compatibility
preview_button(bootstrap("solar@4-3"))

# Set bootstrap SASS variables (globally)
# To set global themes, you start by calling bs_theme_new()
bs_theme_new()
bs_theme_add_variables(
  primary = "orange",
  "body-bg" = "#EEEEEE",
  "font-family-base" = "monospace",
  "font-size-base" = "1.4rem",
  "btn-padding-y" = ".16rem",
  "btn-padding-x" = "2rem",
  "border-radius" = 0,
  "border-radius-lg" = 0,
  "border-radius-sm" = 0
)
preview_button(bootstrap())

# Start a new global theme
bs_theme_new()
# Include custom CSS that leverages bootstrap SASS variables
person <- function(name, title, company) {
  tags$div(
    class = "person",
    h3(class = "name", name),
    div(class = "title", title),
    div(class = "company", company)
  )
}
person_scss <- sass::sass_file(system.file("custom", "person.scss", package = "bootstraplib"))
browsable(tags$body(
  bootstrap(),
  tags$style(bootstrap_sass(person_scss)),
  person("Andrew Carnegie", "Owner", "Carnegie Steel Company"),
  person("John D. Rockefeller", "Chairman", "Standard Oil")
))

}
\references{
\url{https://getbootstrap.com/docs/4.3/getting-started/theming/}
}
\seealso{
\code{\link[=bs_theme_set]{bs_theme_set()}}
}
