% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootswatch.R
\name{bootswatch_themes}
\alias{bootswatch_themes}
\alias{has_bootswatch_theme}
\title{Bootswatch helpers}
\usage{
bootswatch_themes(version = version_latest(), full_path = FALSE)

has_bootswatch_theme(theme)
}
\arguments{
\item{version}{the major version of Bootswatch.}

\item{full_path}{whether to return a path to the installed theme.}
}
\description{
Use \code{bootswatch_themes()} to obtain a list of all available bootswatch themes.
Use \code{is_bootswatch_theme()} to test if a \code{\link[=bs_theme]{bs_theme()}} object has a bootswatch theme.
}
