% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{bootswatch_detect}
\alias{bootswatch_detect}
\title{Detect for Bootswatch themes}
\usage{
bootswatch_detect(x, version = version_latest())
}
\arguments{
\item{x}{SASS code that may contain \link{theme_layer_bootswatch}}

\item{version}{Bootswatch version}
}
\value{
Bootswatch theme names as a character string (if any); otherwise, \code{NA}
(if no Bootswatch themes are detected).
}
\description{
Detect a SASS input for Bootswatch themes.
}
