% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{breakpoints}
\alias{breakpoints}
\title{Define breakpoint values}
\usage{
breakpoints(..., sm = NULL, md = NULL, lg = NULL)
}
\arguments{
\item{...}{Other breakpoints (e.g., \code{xl}).}

\item{sm}{Values to apply at the \code{sm} breakpoint.}

\item{md}{Values to apply at the \code{md} breakpoint.}

\item{lg}{Values to apply at the \code{lg} breakpoint.}
}
\description{
A generic constructor for responsive breakpoints.
}
\references{
\url{https://getbootstrap.com/docs/5.3/layout/breakpoints/}
}
\seealso{
\code{\link[=layout_columns]{layout_columns()}}
}
