% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{breakpoints_column_widths}
\alias{breakpoints_column_widths}
\title{Define responsive breakpoints}
\usage{
breakpoints_column_widths(sm = NULL, md = NULL, lg = NULL, ...)
}
\description{
TODO: describe me
}
