% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs4_sass.R
\name{bs4_sass}
\alias{bs4_sass}
\alias{bs4_scss_file}
\title{Compile Bootstrap 4 SASS with (optional) theming}
\usage{
bs4_sass(sass_vars = sass_variables(), theme = bs4_theme(),
  bootstrap_scss = bs4_scss_file("bootstrap.scss"))

bs4_scss_file(path)
}
\arguments{
\item{sass_vars}{A list of SASS variables to include prior to any bootstrap or
\code{theme} sass. Set to \code{NULL} if you want to ignore any existing variables
(set by \code{\link[=sass_variables_add]{sass_variables_add()}}) at runtime.}

\item{theme}{a \code{\link[=bs4_theme]{bs4_theme()}} object. Use this argument to define your own custom
theme and/or use a pre-defined theme (e.g., \code{\link[=bs4_theme_bootswatch]{bs4_theme_bootswatch()}}).}
}
\description{
For vanilla Bootstrap CSS and all it's dependencies, call \code{bs4_sass()}.
See examples below for various ways to generating custom Bootstrap themes.
}
\examples{
library(htmltools)

# Vanilla Bootstrap 4
button <- tags$a(class = "btn btn-primary", href = "#", role = "button", "Hello")
browsable(tags$body(bs4_sass(), button))

# Use a bootswatch theme
bs4_minty <- bs4_sass(theme = bs4_theme_bootswatch("minty"))
browsable(tags$body(bs4_minty, button))

# Set bootstrap SASS variables (globally)
sass_variables_add(primary = "orange", "body-bg" = "#EEEEEE")
sass_variables_add("font-family-base" = "monospace", "font-size-base" = "1.4rem")
sass_variables_add("btn-padding-y" = ".16rem")
sass_variables_add("btn-padding-x" = "2rem")
sass_variables_add("border-radius" = 0, "border-radius-lg" = 0, "border-radius-sm" = 0)
browsable(tags$body(bs4_sass(), button))

# Use a custom theme
person <- function(name, title, company) {
  tags$div(
    class = "person",
    h3(class = "name", name),
    div(class = "title", title),
    div(class = "company", company)
  )
}
person_scss <- sass::sass_file(system.file("custom", "person.scss", package = "bootsass"))
browsable(tags$body(
  bs4_sass(theme = bs4_theme(post = person_scss)),
  person("Andrew Carnegie", "Owner", "Carnegie Steel Company"),
  person("John D. Rockefeller", "Chairman", "Standard Oil")
))

}
