% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{bs4_theme}
\alias{bs4_theme}
\alias{bs4_themes}
\alias{bs4_theme_bootswatch}
\alias{bs4_theme_bs3compat}
\alias{bootswatch_themes}
\title{Create a bs4 theme}
\usage{
bs4_theme(pre = "", post = "")

bs4_themes(...)

bs4_theme_bootswatch(theme = "")

bs4_theme_bs3compat()

bootswatch_themes(full.names = FALSE)
}
\arguments{
\item{pre}{An acceptable \code{\link[sass:sass]{sass::sass()}} input. This SASS input will be included before bootstrap's SASS}

\item{post}{An acceptable \code{\link[sass:sass]{sass::sass()}} input. This SASS input will be included before bootstrap's SASS}

\item{...}{\code{bs4_theme()} objects.}

\item{theme}{a \url{https://bootswatch.com/} theme name.}
}
\description{
A bs4 theme defines sass input to place before and/or after Bootstrap's SASS.
}
\references{
\url{https://getbootstrap.com/docs/4.3/getting-started/theming/}, \url{https://bootswatch.com/}
}
\seealso{
\code{\link[=bs4_sass]{bs4_sass()}}
}
