% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-theme-layers.R
\name{bs_add_defaults}
\alias{bs_add_defaults}
\alias{bs_add_variables}
\alias{bs_add_layers}
\alias{bs_add_rules}
\alias{bs_add_declarations}
\title{Add low-level theming customizations}
\usage{
bs_add_defaults(theme, ...)

bs_add_variables(
  theme,
  ...,
  .where = "defaults",
  .default_flag = identical(.where, "defaults")
)

bs_add_layers(theme, ...)

bs_add_rules(theme, rules)

bs_add_declarations(theme, declarations)
}
\arguments{
\item{theme}{a \code{\link[=bs_theme]{bs_theme()}} object.}

\item{...}{For \code{bs_add_variables()}, these arguments should contain Sass
variables; otherwise, these arguments should contain \code{\link[sass:sass_layer]{sass::sass_layer()}}s.}

\item{.where}{whether to place the variable definitions before other Sass
\code{"defaults"}, after other Sass \code{"declarations"}, or after other Sass
\code{"rules"}.}

\item{.default_flag}{whether or not to add a \code{!default} flag (if missing) to
variable expressions. It's recommended to keep this as \code{TRUE} when \code{.where = "defaults"}.}

\item{rules}{Sass rules.}

\item{declarations}{Sass functions and mixins.}
}
\description{
Use \code{\link[=bs_add_variables]{bs_add_variables()}} to add Bootstrap Sass \href{https://getbootstrap.com/docs/4.4/getting-started/theming/#variable-defaults}{variable defaults},
\code{\link[=bs_add_rules]{bs_add_rules()}} to add additional \href{https://sass-lang.com/documentation/style-rules}{Sass rules}, and
\code{\link[=bs_add_declarations]{bs_add_declarations()}} to add Sass
\href{https://sass-lang.com/documentation/at-rules/function}{functions} and
\href{https://sass-lang.com/documentation/at-rules/mixin}{mixins}. Compared to
higher-level theme customization available in \code{\link[=bs_theme]{bs_theme()}}, these functions
are a more direct interface to Bootstrap Sass, and therefore, do nothing to
ensure the theme customization you add are portable between major Bootstrap
versions.
}
\examples{

# Function to preview the styling a (primary) Bootstrap button
library(htmltools)
button <- tags$a(class = "btn btn-primary", href = "#", role = "button", "Hello")
preview_button <- function(theme) {
  browsable(tags$body(bs_theme_dependencies(theme), button))
}

# Here we start with a theme based on a Bootswatch theme,
# then override some variable defaults
theme <- bs_theme(bootswatch = "sketchy", primary = "orange") \%>\%
  bs_add_variables(
    "body-bg" = "#EEEEEE",
    "font-family-base" = "monospace",
    "font-size-base" = "1.4rem",
    "btn-padding-y" = ".16rem",
    "btn-padding-x" = "2rem",
    "border-radius" = 0,
    "border-radius-lg" = 0,
    "border-radius-sm" = 0
  )

preview_button(theme)

# If you need to set a variable based on another Bootstrap variable
theme \%>\%
  bs_add_variables("body-color" = "$success", .where = "declarations") \%>\%
  preview_button()

# Start a new global theme and add some custom rules that
# use Bootstrap variables to define a custom styling for a
# 'person card'
person_rules <- system.file("custom", "person.scss", package = "bootstraplib")
theme <- bs_add_rules(bs_theme(), rules = sass::sass_file(person_rules))
# Include custom CSS that leverages bootstrap Sass variables
person <- function(name, title, company) {
  tags$div(
    class = "person",
    h3(class = "name", name),
    div(class = "title", title),
    div(class = "company", company)
  )
}
browsable(tags$body(
  bs_theme_dependencies(theme),
  person("Andrew Carnegie", "Owner", "Carnegie Steel Company"),
  person("John D. Rockefeller", "Chairman", "Standard Oil")
))

}
\references{
\url{https://getbootstrap.com/docs/4.4/getting-started/theming/}

\url{https://rstudio.github.io/sass/articles/sass.html#layering}
}
