% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{bs_card}
\alias{bs_card}
\title{Create a Bootstrap Card Component}
\usage{
bs_card(
  ...,
  header = NULL,
  footer = NULL,
  lead = NULL,
  height = NULL,
  width = NULL,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{...}{UI elements for the \code{card()}'s body.}

\item{header}{A character string or UI element for the header.}

\item{footer}{A character string or UI element for the footer.}

\item{lead}{A character string or UI element for the body's main text.}

\item{title}{A character string or UI element for the body's title.}

\item{subtitle}{A character string or UI element for the body's subtitle.}
}
\value{
The newly created card container.
}
\description{
A wrapper for Bootstrap 4+'s \code{card} component.
}
\examples{

library(htmltools)

card1 <- bs_card("This is some text within a card body.")

card2 <- bs_card(
  header = "Featured",
  title = "Special title treatment",
  lead = "With supporting text below as a natural lead-in to additional content.",
  tags$a("Go somewhere", href = "#", class = "btn btn-primary")
)

bs_cards_group(card1, card2)
bs_cards_grid(card1, card2)
bs_cards_grid(card1, card2, equal_height = TRUE)

bs_card(
  header = "Look, a cool plot!",
  plotTag(image(volcano), "A volcano")
)


bs_card(
  title = "Look, a cool plot!",
  plotTag(image(volcano), "A volcano")
)
}
\concept{Component}
