% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-current-theme.R
\name{bs_current_theme}
\alias{bs_current_theme}
\title{Obtain the currently active theme}
\usage{
bs_current_theme(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{}
}
\value{
a \code{\link[=bs_theme]{bs_theme()}} object.
}
\description{
Obtain the currently active theme
}
