% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-dependencies.R
\name{bs_dependencies}
\alias{bs_dependencies}
\alias{bs_sass}
\title{Compile Bootstrap SASS with (optional) theming}
\usage{
bs_dependencies(
  theme,
  sass_options = sass::sass_options(output_style = "compressed"),
  cache = sass::sass_cache_get(),
  jquery = jquerylib::jquery_core(3),
  precompiled = TRUE
)

bs_sass(rules = list(), theme, write_attachments = FALSE, ...)
}
\arguments{
\item{theme}{a \code{\link[=bs_theme]{bs_theme()}} object.}

\item{sass_options}{a \code{\link[sass:sass_options]{sass::sass_options()}} object.}

\item{cache}{This can be a directory to use for the cache, a \link[sass]{FileCache}
object created by \code{\link[sass:sass_file_cache]{sass_file_cache()}}, or \code{FALSE} or \code{NULL} for no caching.}

\item{jquery}{a \code{\link[jquerylib:jquery_core]{jquerylib::jquery_core()}} object.}

\item{precompiled}{Before compiling the theme object, first look for a
precompiled CSS file for the given \code{version}.  If this option is \code{TRUE} and
a precompiled CSS file exists for the theme object, it will be fetched
immediately and not compiled. At the moment, we only provide precompiled
CSS for "stock" builds of Bootstrap (i.e., no theming additions, bootswatch
themes, or non-default \code{sass_options}).}

\item{rules}{Sass styling rules that may reference the \code{theme}'s \code{defaults} and \code{declarations}.}

\item{write_attachments}{If the input contains \code{\link[sass:sass_layer]{sass_layer()}} objects that
have file attachments, and \code{output} is not \code{NULL}, then copy the file
attachments to the directory of \code{output}. (Defaults to \code{NA}, which merely
emits a warning if file attachments are present, but does not write them to
disk; the side-effect of writing extra files is subtle and potentially
destructive, as files may be overwritten.)}

\item{...}{arguments passed along to \code{\link[sass:sass]{sass::sass()}}.}
}
\value{
a list of HTML dependencies containing Bootstrap CSS, Bootstrap
JavaScript, and \code{jquery}. This list may contain additional HTML
dependencies if the \code{theme} calls for it (e.g., \code{version = "4+3"} contains
an additional JavaScript dependency).
}
\description{
Use \code{bs_dependencies()} to compile Bootstrap Sass into CSS and return it, along
with other HTML dependencies, as a list of \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}s. Use
\code{bs_sass()} if you can assume Bootstrap already exists on the page,
but you want to leverage Bootstrap utilities (e.g., variables, functions, or
mixins) to generate additional CSS rules (as a string that can be included as
a \verb{<style>} tag via \code{tags$style(css)}).
}
\examples{

# Function to preview the styling a (primary) Bootstrap button
library(htmltools)
button <- tags$a(class = "btn btn-primary", href = "#", role = "button", "Hello")
preview_button <- function(x) {
  browsable(tags$body(x, button))
}

# Latest bootstrap
preview_button(bs_dependencies())
# Bootstrap 3
preview_button(bs_dependencies("3"))
# Bootswatch minty theme
preview_button(bs_dependencies("minty"))
# Bootswatch sketchy theme
preview_button(bs_dependencies("sketchy"))
# Bootswatch solar theme with BS3 compatibility
preview_button(bs_dependencies("solar@4+3"))

}
\seealso{
\code{\link[=bs_theme]{bs_theme()}}, \code{\link[=bs_global_set]{bs_global_set()}}
}
