% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_sass.R
\name{bs_sass}
\alias{bs_sass}
\alias{bs4_scss_required}
\alias{bs4_scss_file}
\title{Compile Bootstrap 4 SASS with (optional) theming}
\usage{
bs_sass(
  ...,
  variables = theme_variables(),
  bootstrap_scss = bs4_scss_file("bootstrap.scss"),
  bs3compat = TRUE,
  options = sass_options()
)

bs4_scss_required()

bs4_scss_file(path)
}
\arguments{
\item{...}{a collection of \code{\link[sass:sass]{sass::sass()}} input and/or \code{\link[=bs_themes]{bs_themes()}}.
Use this argument to supply pre-defined theme(s) like \code{\link[=theme_bootswatch]{theme_bootswatch()}} and
combine those theme with your own SASS code.}

\item{variables}{A list of SASS variables to include prior to any other SASS.
To ignore any existing variables (set by \code{\link[=theme_variables]{theme_variables()}}) at runtime,
set this to \code{NULL}.}

\item{bootstrap_scss}{bootstrap scss files to compile \code{variables} and \code{...} against.
By default, all of bootstrap is included. Use \code{bs4_scss_required()} to compile against
'core' parts of Bootstrap 4. If you need to use a specific version of Bootstrap,
supply it via \code{\link[sass:sass_file]{sass::sass_file()}} and attach a \code{"bootstrap_version"} attribute
for the version to be reflected in the returned HTML dependency.}

\item{options}{Compiler options for SASS, see \code{\link[sass:sass_options]{sass::sass_options()}}.}
}
\value{
a list of HTML dependencies.
}
\description{
Obtain Bootstrap (and it's dependencies) as an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
By default, \code{bs_sass()} returns vanilla Bootstrap, but it also provides
various ways to theme (see examples below).
}
\examples{
library(htmltools)

# Vanilla Bootstrap 4
button <- tags$a(class = "btn btn-primary", href = "#", role = "button", "Hello")
browsable(tags$body(bs_sass(), button))

# Use a bootswatch theme
minty <- bs_sass(theme_bootswatch("minty"))
browsable(tags$body(minty, button))

# Set bootstrap SASS variables (globally)
theme_variables(primary = "orange", "body-bg" = "#EEEEEE")
theme_variables("font-family-base" = "monospace", "font-size-base" = "1.4rem")
theme_variables("btn-padding-y" = ".16rem")
theme_variables("btn-padding-x" = "2rem")
theme_variables("border-radius" = 0, "border-radius-lg" = 0, "border-radius-sm" = 0)
browsable(tags$body(bs_sass(), button))

# Use a custom theme
person <- function(name, title, company) {
  tags$div(
    class = "person",
    h3(class = "name", name),
    div(class = "title", title),
    div(class = "company", company)
  )
}
person_scss <- sass::sass_file(system.file("custom", "person.scss", package = "bootsass"))
browsable(tags$body(
  bs_sass(person_scss),
  person("Andrew Carnegie", "Owner", "Carnegie Steel Company"),
  person("John D. Rockefeller", "Chairman", "Standard Oil")
))

}
\references{
\url{https://getbootstrap.com/docs/4.3/getting-started/theming/}
}
\seealso{
\code{\link[=bs_themes]{bs_themes()}}, \code{\link[=theme_variables]{theme_variables()}}
}
