% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-theme.R
\name{bs_theme}
\alias{bs_theme}
\alias{bs_theme_update}
\alias{bs_global_theme_update}
\title{Create (and update) a Bootstrap theme object}
\usage{
bs_theme(
  version = version_default(),
  bootswatch = NULL,
  ...,
  bg = NULL,
  fg = NULL,
  primary = NULL,
  secondary = NULL,
  success = NULL,
  info = NULL,
  warning = NULL,
  danger = NULL,
  base_font = NULL,
  code_font = NULL,
  heading_font = NULL
)

bs_theme_update(
  theme = bs_theme(),
  ...,
  bg = NULL,
  fg = NULL,
  primary = NULL,
  secondary = NULL,
  success = NULL,
  info = NULL,
  warning = NULL,
  danger = NULL,
  base_font = NULL,
  code_font = NULL,
  heading_font = NULL
)

bs_global_theme_update(
  theme = bs_theme(),
  ...,
  bg = NULL,
  fg = NULL,
  primary = NULL,
  secondary = NULL,
  success = NULL,
  info = NULL,
  warning = NULL,
  danger = NULL,
  base_font = NULL,
  code_font = NULL,
  heading_font = NULL
)
}
\arguments{
\item{version}{The major version of Bootstrap to use. A value of
\code{'4+3'} means Bootstrap 4, but with additional CSS/JS to support
BS3 style markup in BS4. Other supported versions include 3 and 4.}

\item{bootswatch}{The name of a bootswatch theme (see \code{\link[=bootswatch_themes]{bootswatch_themes()}} for possible values).}

\item{...}{arguments passed along to \code{\link[=bs_add_variables]{bs_add_variables()}}.}

\item{bg}{A color string for the background, in any format \code{\link[htmltools:parseCssColors]{htmltools::parseCssColors()}} can understand.}

\item{fg}{A color string for the foreground.}

\item{primary}{A color to be used for hyperlinks, to indicate primary/default
actions, and to show active selection state in some Bootstrap components.
Generally a bold, saturated color that contrasts with the theme's base
colors.}

\item{secondary}{A color for components and messages that don't need to stand
out. (Not supported in Bootstrap 3.)}

\item{success}{A color for messages that indicate an operation has succeeded.
Typically green.}

\item{info}{A color for messages that are informative but not critical. Typically a
shade of blue-green.}

\item{warning}{A color for warning messages. Typically yellow.}

\item{danger}{A color for errors. Typically red.}

\item{base_font}{The default typeface.}

\item{code_font}{The typeface to be used for code. Be sure this is monospace!}

\item{heading_font}{The typeface to be used for heading elements.}

\item{theme}{a \code{\link[=bs_theme]{bs_theme()}} object.}
}
\description{
Create (and update) a Bootstrap theme object
}
\section{Main colors}{

\subsection{Grayscale colors}{

Specify both \code{fg} (foreground) and \code{bg} (background) colors to control
the default grayscale palette used to set the vast majority of color defaults.
}

\subsection{Accent colors:}{
}
}

\section{Fonts}{


Set the typefaces used by Bootstrap for various purposes. Each argument
can be \code{NULL} (no change), or a character vector of one or more elements.

Each argument is a character vector, and each element of that vector can a
single unquoted font family name, a single quoted font family name, or a
comma-separated list of font families (with individual font family names
quoted as necessary).

For example, each example below is valid:\preformatted{# Single, unquoted
bs_theme_fonts(base = "Source Sans Pro")

# Single, quoted
bs_theme_fonts(base = "'Source Sans Pro'")

# Multiple, quoted
bs_theme_fonts(base = "'Source Sans Pro', sans-serif")

# Combining all of the above
bs_theme_fonts(base = c("Open Sans", "'Source Sans Pro'",
  "'Helvetica Neue', Helvetica, sans-serif"))
}

But the following is \emph{technically} not valid:\preformatted{# Incorrect--because multiple font families are being
# provided in a single string, names with spaces must
# be surrounded by quotes!
bs_theme_fonts(base = "Source Sans Pro, sans-serif")
}

The resulting CSS will contain \verb{font-family: Source Sans Pro, sans-serif;}
which is technically out of spec, but in fact is likely to still work with
most browsers.
}

\section{Miscellaneous Sass variables}{


Both Bootstrap 3 and 4 come with

TODO: link to the variables files?
}

\examples{

theme <- bs_theme(
  # Controls the default grayscale palette
  bg = "#202123", fg = "#B8BCC2",
  # Controls the accent (e.g., hyperlink, button, etc) colors
  primary = "#EA80FC", secondary = "#48DAC6",
  base_font = c("Grandstander", "sans-serif"),
  code_font = c("Courier", "monospace"),
  heading_font = "'Helvetica Neue', Helvetica, sans-serif",
  # Can also add lower-level customization
  "input-border-color" = "#EA80FC"
)
if (interactive()) {
  bs_theme_preview(theme)
}

# Lower-level bs_add_*() functions allow you to work more
# directly with the underlying Sass code
theme <- theme \%>\%
  bs_add_variables("my-class-color" = "red") \%>\%
  bs_add_rules(".my-class { color: $my-class-color }")

}
\references{
\url{https://getbootstrap.com/docs/4.4/getting-started/theming/}

\url{https://rstudio.github.io/sass/}
}
