% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-base-colors.R
\name{bs_theme_accent_colors}
\alias{bs_theme_accent_colors}
\title{Customize Bootstrap accent colors}
\usage{
bs_theme_accent_colors(
  primary = NULL,
  secondary = NULL,
  success = NULL,
  info = NULL,
  warning = NULL,
  danger = NULL
)
}
\arguments{
\item{primary}{A color to be used for hyperlinks, to indicate primary/default
actions, and to show active selection state in some Bootstrap components.
Generally a bold, saturated color that contrasts with the theme's base
colors.}

\item{secondary}{A color for components and messages that don't need to stand
out. (Not supported in Bootstrap 3.)}

\item{success}{A color for messages that indicate an operation has succeeded.
Typically green.}

\item{info}{A color for messages that are informative but not critical. Typically a
shade of blue-green.}

\item{warning}{A color for warning messages. Typically yellow.}

\item{danger}{A color for errors. Typically red.}
}
\description{
Set accent colors (referred to as "brand" colors in the Bootstrap 3 docs, and
"theme" colors in the Bootstrap 4 docs) for the current Bootstrap theme.
Values must be \code{NULL} or a color string in a format
\code{\link[htmltools:parseCssColors]{htmltools::parseCssColors()}} can understand.
}
\examples{
bs_theme_new("4+3")
bs_theme_accent_colors(primary = "maroon", secondary = "gray")
if (interactive()) {
  bs_theme_preview(with_themer = FALSE)
}

}
\seealso{
Other customizations: 
\code{\link{bs_theme_base_colors}()},
\code{\link{bs_theme_fonts}()}
}
\concept{customizations}
