% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick-theme.R
\name{bs_theme_base_colors}
\alias{bs_theme_base_colors}
\title{Customize the Bootstrap theme based on two to four key colors}
\usage{
bs_theme_base_colors(
  bg = "#FFFFFF",
  fg = "#000000",
  accent = NULL,
  secondary = NULL
)
}
\arguments{
\item{bg}{A color string for the background, in any format
\code{\link[htmltools:parseCssColors]{htmltools::parseCssColors()}} can understand.}

\item{fg}{A color string for the background.}

\item{accent}{A color string for the accent color; if not \code{NULL}, this will
be assigned to the \verb{$primary} (BS4) or \verb{$brand-primary} (BS3) Sass
variables.}

\item{secondary}{A color string for the secondary color; if not \code{NULL}, this
will be assigned to the \verb{$secondary} and \verb{$default} Bootstrap 4 Sass
variables. (This argument is not currently supported for Bootstrap 3.)}
}
\description{
\code{bs_theme_base_colors} generates a complete Bootstrap theme from just a small
handful of user-specified colors: a background color, a foreground color, and
optionally, primary and secondary accent colors.
}
\section{Implementation notes}{


\code{bs_theme_base_colors} supports both Bootstrap 3 and 4 (and 4+3), but the
implementation differs greatly.

For Bootstrap 4, the vast majority of default colors are directly or
indirectly based on the \verb{$black}, \verb{$white}, and \verb{$gray-100} through
\verb{$gray-900} variables; or on the theme colors (primary, secondary, danger,
warning, info, etc.). \code{bs_theme_base_colors} sets \verb{$white} to the \code{bg} color,
\verb{$black} to the \code{fg} color, and interpolates the grays between them. If
provided, the \code{accent} argument is used to set the \verb{$primary} variable, and
the \code{secondary} argument is used to set \verb{$secondary} and \verb{$default}
variables.

For Bootstrap 3, a similar set of \verb{$black}, \verb{$white}, and \verb{$gray-darker}
through \verb{$gray-lighter} variables exist, and these are populated using the
same strategy as with Bootstrap 4. However, unlike Bootstrap 4, in Bootstrap
3 many of the default colors are hard-coded hex colors, that also happen to
be shades of gray. \code{bs_theme_base_colors} overrides these hard-coded values
with colors interpolated between \code{bg} and \code{fg}.
}

\examples{

bs_theme_new("4+3")
bs_theme_base_colors(bg = "#000060", fg = "skyblue",
  accent = "orange", secondary = "silver")

# You can apply further customizations here if desired, e.g.:
bs_theme_add_variables("success" = "#1D7732")

if (interactive()) {
  bs_theme_preview()
}

}
