% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-dependencies.R
\name{bs_theme_dependencies}
\alias{bs_theme_dependencies}
\title{Compile Bootstrap Sass with (optional) theming}
\usage{
bs_theme_dependencies(
  theme,
  sass_options = sass::sass_options(output_style = "compressed"),
  cache = sass::sass_cache_get(),
  jquery = jquerylib::jquery_core(3),
  precompiled = TRUE
)
}
\arguments{
\item{theme}{a \code{\link[=bs_theme]{bs_theme()}} object.}

\item{sass_options}{a \code{\link[sass:sass_options]{sass::sass_options()}} object.}

\item{cache}{This can be a directory to use for the cache, a \link[sass]{FileCache}
object created by \code{\link[sass:sass_file_cache]{sass_file_cache()}}, or \code{FALSE} or \code{NULL} for no caching.}

\item{jquery}{a \code{\link[jquerylib:jquery_core]{jquerylib::jquery_core()}} object.}

\item{precompiled}{Before compiling the theme object, first look for a
precompiled CSS file for the given \code{version}.  If this option is \code{TRUE} and
a precompiled CSS file exists for the theme object, it will be fetched
immediately and not compiled. At the moment, we only provide precompiled
CSS for "stock" builds of Bootstrap (i.e., no theming additions, bootswatch
themes, or non-default \code{sass_options}).}
}
\value{
a list of HTML dependencies containing Bootstrap CSS, Bootstrap
JavaScript, and \code{jquery}. This list may contain additional HTML
dependencies if the \code{theme} calls for it (e.g., \code{version = "4+3"} contains
an additional JavaScript dependency).
}
\description{
Use \code{bs_theme_dependencies()} to compile Bootstrap Sass into CSS and return it, along
with other HTML dependencies, as a list of \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}s. Use
\code{sass_partial()} if you can assume Bootstrap already exists on the page,
but you want to leverage Bootstrap utilities (e.g., variables, functions, or
mixins) to generate additional CSS rules (as a string that can be included as
a \verb{<style>} tag via \code{tags$style(css)}).
}
\examples{

# Function to preview the styling a (primary) Bootstrap button
library(htmltools)
button <- tags$a(class = "btn btn-primary", href = "#", role = "button", "Hello")
preview_button <- function(theme) {
  theme \%>\%
    bs_theme_dependencies() \%>\%
    tags$body(button) \%>\%
    browsable()
}

# Latest bootstrap
preview_button(bs_theme())
# Bootstrap 3
preview_button(bs_theme(3))
# Bootswatch minty theme
preview_button(bs_theme(bootswatch = "minty"))
# Bootswatch sketchy theme
preview_button(bs_theme(bootswatch = "sketchy"))
# Bootswatch solar theme with BS3 compatibility
preview_button(bs_theme(version = "4+3", bootswatch = "solar"))

}
\seealso{
\code{\link[=bs_theme]{bs_theme()}}, \code{\link[=bs_global_set]{bs_global_set()}}
}
