% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-base-colors.R
\name{bs_theme_fonts}
\alias{bs_theme_fonts}
\title{Customize Bootstrap typefaces}
\usage{
bs_theme_fonts(base = NULL, code = NULL, heading = NULL)
}
\arguments{
\item{base}{The default typeface.}

\item{code}{The typeface to be used for code. Be sure this is monospace!}

\item{heading}{The typeface to be used for heading elements.}
}
\description{
Set the typefaces used by Bootstrap for various purposes. Each argument
can be \code{NULL} (no change), or a character vector of one or more elements.
}
\details{
Each argument is a character vector, and each element of that vector can
either be a font family name, a CSS-quoted or -escaped font family name,
or a comma-separated list of CSS-quoted or -escaped font family names.
See the example below for examples of all three.
}
\examples{

bs_theme_new()
bs_theme_fonts(
  base = "Times",
  code = c("Courier", "monospace"),
  heading = "'Helvetica Neue', Helvetica, sans-serif"
)

if (interactive()) {
  bs_theme_preview(with_themer = FALSE)
}

}
