% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-preview.R
\name{bs_theme_get_variables}
\alias{bs_theme_get_variables}
\title{Retrieve Sass variable values from the current theme}
\usage{
bs_theme_get_variables(varnames)
}
\arguments{
\item{varnames}{a character string referencing a Sass variable
in the current theme.}
}
\value{
a character string containing a CSS/Sass value.
If the variable(s) are not defined, their value is \code{NA}.
}
\description{
Useful for retriving a variable from the current theme and using
the value to inform another R function.
}
\examples{
bs4_vars <- c("body-bg", "body-color", "primary", "border-radius")
bs_theme_new()
bs_theme_get_variables(bs4_vars)
bs_theme_new(bootswatch = "darkly")
bs_theme_get_variables(bs4_vars)

}
