% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-preview.R
\name{bs_theme_preview}
\alias{bs_theme_preview}
\title{Preview the currently set theme}
\usage{
bs_theme_preview()
}
\description{
Launches an example shiny app via \code{run_with_themer()} and \code{bootstrap()}.
Useful for getting a quick preview of the current theme setting as
well as an interactive GUI for tweaking some of the main theme settings.
}
\details{
The app that this launches is subject to change.
}
\examples{

bs_theme_new()
bs_theme_add_variables(
  "body-bg" = "#6c757d",
  "body-color" = "white"
)
if (interactive()) {
  bs_theme_preview()
}

}
\seealso{
\code{\link[=run_with_themer]{run_with_themer()}}
}
