% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{bs_themes}
\alias{bs_themes}
\alias{bs_theme}
\alias{theme_bootswatch}
\alias{theme_bs3compat}
\alias{bootswatch_themes}
\title{Pre-defined and custom Bootstrap 4 themes}
\usage{
bs_themes(...)

bs_theme(pre = "", post = "", deps = NULL)

theme_bootswatch(theme = "")

theme_bs3compat()

bootswatch_themes(full.names = FALSE)
}
\arguments{
\item{...}{A collection of \code{\link[sass:sass_layer]{sass_layer()}}s and/or objects that \code{\link[sass:as_sass]{as_sass()}} understands.}

\item{pre}{An object that \code{\link[sass:as_sass]{as_sass()}} understands.}

\item{post}{An object that \code{\link[sass:as_sass]{as_sass()}} understands.}

\item{deps}{An HTML dependency (or a list of them).}

\item{theme}{a \url{https://bootswatch.com/} theme name.}
}
\description{
Pre-defined and custom Bootstrap 4 themes
}
