% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bslib-package.R
\docType{package}
\name{bslib-package}
\alias{bslib}
\alias{bslib-package}
\title{bslib: Custom 'Bootstrap' 'Sass' Themes for 'shiny' and 'rmarkdown'}
\description{
Simplifies custom 'CSS' styling of both 'shiny' and 'rmarkdown' via 'Bootstrap' 'Sass'. Supports 'Bootstrap' 3, 4 and 5 as well as their various 'Bootswatch' themes. An interactive widget is also provided for previewing themes in real time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/bslib/}
  \item \url{https://github.com/rstudio/bslib}
  \item Report bugs at \url{https://github.com/rstudio/bslib/issues}
}

}
\author{
\strong{Maintainer}: Carson Sievert \email{carson@posit.co} (\href{https://orcid.org/0000-0002-4958-2844}{ORCID})

Authors:
\itemize{
  \item Joe Cheng \email{joe@posit.co}
  \item Garrick Aden-Buie \email{garrick@posit.co} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item  Bootstrap contributors (Bootstrap library) [contributor]
  \item  Twitter, Inc (Bootstrap library) [copyright holder]
  \item Javi Aguilar (Bootstrap colorpicker library) [contributor, copyright holder]
  \item Thomas Park (Bootswatch library) [contributor, copyright holder]
  \item  PayPal (Bootstrap accessibility plugin) [contributor, copyright holder]
}

}
\keyword{internal}
