% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-theme-preset-builtin.R
\name{builtin_themes}
\alias{builtin_themes}
\title{Obtain a list of all available built-in \pkg{bslib} themes.}
\usage{
builtin_themes(version = version_default(), full_path = FALSE)
}
\arguments{
\item{version}{the major version of Bootstrap.}

\item{full_path}{whether to return a path to the installed theme.}
}
\value{
a character vector of built-in themes provided by \pkg{bslib}.
}
\description{
Obtain a list of all available built-in \pkg{bslib} themes.
}
