% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{as.card_item}
\alias{as.card_item}
\alias{is.card_item}
\alias{card_body}
\alias{card_body_scroll}
\alias{card_header}
\alias{card_title}
\alias{card_footer}
\alias{card_spacer}
\alias{card_body_output}
\alias{card_body_grid}
\title{Card item components}
\usage{
as.card_item(x)

is.card_item(x)

card_body(..., stretch = FALSE, class = NULL, padding = c("x", "y"))

card_body_scroll(..., height = NULL, class = NULL, padding = c("x", "y"))

card_header(..., class = NULL, container = htmltools::div)

card_title(..., class = NULL, padding = c("x", "y"), container = htmltools::h5)

card_footer(..., class = NULL)

card_spacer(...)

card_body_output(output, stretch = TRUE, height = if (!stretch) "400px", ...)

card_body_grid(
  ...,
  width = 1/4,
  gap = NULL,
  fixed_width = FALSE,
  stretch = TRUE,
  padding = c("x", "y")
)
}
\arguments{
\item{x}{an object to test (or coerce to) a card item.}

\item{...}{Named arguments become attributes on the \verb{<div class="card">}
element. Unnamed arguments become card items, and can be any valid child of
an \link[htmltools:builder]{htmltools tag}.}

\item{stretch}{Set to \code{TRUE} if this \code{card_body} is eager to use any extra
vertical space is available in the card.}

\item{class}{Additional CSS classes to include on the card div.}

\item{padding}{Whether vertical and/or horizontal padding should be included.
Use \code{NULL} for neither, \code{"x"} for horizontal only, \code{"y"} for vertical only,
or \code{c("x", "y")} for both (the default).}

\item{height}{Any valid \link[htmltools:validateCssUnit]{CSS unit}; for
example, \code{height="100\%"}.}

\item{container}{a function to generate an HTML element.}

\item{output}{the result of a \verb{*Output()} function (e.g, \code{\link[shiny:plotOutput]{shiny::plotOutput()}}, \code{plotly::plotlyOutput()}, etc).}
}
\description{
This topic describes various components that are intended to go
directly inside of a \code{\link[=card]{card()}}. These components can be used in combination;
for example, a single card could contain a \code{card_header()}, multiple
\code{card_body()}s, a \code{card_list()}, and finally a \code{card_footer()}.

To learn more about \code{\link[=card]{card()}}s and card layout options, see \href{https://rstudio.github.io/bslib/articles/cards.html}{this article}.
}
\seealso{
\code{\link[=card]{card()}} for creating a card component.

\code{\link[=card_grid]{card_grid()}} for laying out multiple cards.

\code{\link[=navs_tab_card]{navs_tab_card()}} \code{\link[=navs_pill_card]{navs_pill_card()}} for placing navigation in cards.
}
