% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card_grid}
\alias{card_grid}
\title{Card grid}
\usage{
card_grid(
  ...,
  class = NULL,
  card_width = 1/4,
  gap = NULL,
  heights_equal = c("all", "row"),
  fixed_width = FALSE
)
}
\arguments{
\item{...}{Named arguments become attributes on the \verb{<div class="card">}
element. Unnamed arguments are children, and should be instances of
\code{\link[=card]{card()}}.}

\item{class}{Additional CSS classes to include on the grid div.}

\item{card_width}{The desired width of each card, which can be any of the
following:
\itemize{
\item A (unit-less) number between 0 and 1.
\itemize{
\item This should be specified as \code{1/num}, where \code{num} represents the number
of desired columns.
}
\item A \link[htmltools:validateCssUnit]{CSS length unit}
\itemize{
\item Either the minimum (when \code{fixed_width=FALSE}) or fixed width
(\code{fixed_width=TRUE}).
}
\item \code{NULL}
\itemize{
\item Allows power users to set the \code{grid-template-columns} CSS property
manually, either via a \code{style} attribute or a CSS stylesheet.
}
}}

\item{gap}{A \link[htmltools:validateCssUnit]{CSS length unit} that sets the
amount of spacing between cards. If \code{NULL}, the value of Bootstrap's
\verb{$spacer} Sass variable is used.}

\item{heights_equal}{If \code{"all"} (the default), every card in every row of the
grid will have the same height. If \code{"row"}, then every card in \emph{each} row
of the grid will have the same height, but heights may vary between rows.}

\item{fixed_width}{Whether or not to interpret the \code{card_width} as a minimum
(\code{fixed_width=FALSE}) or fixed (\code{fixed_width=TRUE}) width when it is a CSS
length unit.}
}
\description{
Use to create a grid of cards, giving every card in the entire grid the same
height/width (or optionally, every card in \emph{each row} of the grid).
}
