% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-summary.R
\name{card_summary}
\alias{card_summary}
\title{A card for displaying a summary of information}
\usage{
card_summary(
  ...,
  icon = NULL,
  icon_layout = c("top-right", "left-center"),
  class = NULL,
  width = NULL,
  height = NULL,
  color = NULL
)
}
\arguments{
\item{...}{arguments passed along to \code{\link[=card]{card()}}.}

\item{icon}{an icon like \code{\link[shiny:icon]{shiny::icon()}}.}

\item{icon_layout}{where to display the icon (relative to the summary information).}

\item{class}{utility classes for customizing the appearance of the summary card. Use \verb{bg-*} and \verb{text-*} classes (e.g, \code{"bg-danger"} and \code{"text-light"}) to customize the background/foreground colors.}

\item{width, height}{Any valid CSS unit; for example, height="100\%".}

\item{opacity}{amount of opacity (between 0 and 1) to apply to the background color.}
}
\description{
A card for displaying a summary of information
}
\examples{

card_summary(
  "KPI Title",
  card_title("KPI Output"),
  span(icon("arrow-down"), " 30\% VS PREVIOUS 30 DAYS"),
  icon = icon("fire-alt")
)
}
