% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{container}
\alias{container}
\title{Contain, pad, and align content}
\usage{
container(
  ...,
  size = c("sm", "md", "lg", "xl", "xxl", "fluid"),
  bg = NULL,
  class = NULL
)
}
\arguments{
\item{...}{A collection of \code{\link[htmltools:builder]{htmltools::tag()}} children.}

\item{size}{A size (i.e., max-width policy) for the container.}

\item{bg}{A background color.}

\item{class}{Additional CSS classes for the container.}
}
\description{
Contain, pad, and align content
}
\references{
\url{https://getbootstrap.com/docs/5.3/layout/containers/}
}
