% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{bs_theme_new}
\alias{bs_theme_new}
\alias{bs_theme_clear}
\alias{bs_theme_get}
\alias{bs_theme_set}
\alias{bs_theme_base_colors}
\alias{bs_theme_accent_colors}
\alias{bs_theme_fonts}
\alias{bs_theme_add_variables}
\alias{bs_theme_add}
\alias{bs_theme_get_variables}
\alias{bootstrap}
\alias{bootstrap_sass}
\alias{bs_add_declarations}
\alias{card_body_fill}
\alias{page_fill}
\alias{nav}
\alias{nav_content}
\alias{navs_tab}
\alias{navs_pill}
\alias{navs_pill_list}
\alias{navs_hidden}
\alias{navs_bar}
\alias{navs_tab_card}
\alias{navs_pill_card}
\title{Deprecated functions}
\usage{
bs_theme_new(...)

bs_theme_clear(...)

bs_theme_get(...)

bs_theme_set(...)

bs_theme_base_colors(...)

bs_theme_accent_colors(...)

bs_theme_fonts(...)

bs_theme_add_variables(...)

bs_theme_add(...)

bs_theme_get_variables(...)

bootstrap(theme = bs_theme_get(), ...)

bootstrap_sass(rules = list(), theme = bs_theme_get(), ...)

bs_add_declarations(theme, declarations)

card_body_fill(...)

page_fill(...)

nav(...)

nav_content(...)

navs_tab(...)

navs_pill(...)

navs_pill_list(...)

navs_hidden(...)

navs_bar(...)

navs_tab_card(...)

navs_pill_card(...)
}
\value{
a \code{\link[=bs_theme]{bs_theme()}} object.
}
\description{
These functions have been deprecated but remain for backwards compatibility.
}
\section{Functions}{
\itemize{
\item \code{nav()}: Deprecated in v0.5.0, use \code{\link[=nav_panel]{nav_panel()}}.

\item \code{nav_content()}: Deprecated in v0.5.0, use \code{\link[=nav_panel_hidden]{nav_panel_hidden()}}.

\item \code{navs_tab()}: Deprecated in v0.5.0, use \code{\link[=navset_tab]{navset_tab()}}.

\item \code{navs_pill()}: Deprecated in v0.5.0, use \code{\link[=navset_pill]{navset_pill()}}.

\item \code{navs_pill_list()}: Deprecated in v0.5.0, use \code{\link[=navset_pill_list]{navset_pill_list()}}.

\item \code{navs_hidden()}: Deprecated in v0.5.0, use \code{\link[=navset_hidden]{navset_hidden()}}.

\item \code{navs_bar()}: Deprecated in v0.5.0, use \code{\link[=navset_bar]{navset_bar()}}.

\item \code{navs_tab_card()}: Deprecated in v0.5.0, use \code{\link[=navset_card_tab]{navset_card_tab()}}.

\item \code{navs_pill_card()}: Deprecated in v0.5.0, use \code{\link[=navset_card_pill]{navset_card_pill()}}.

}}
\keyword{internal}
