% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-download_button.R
\name{download_button}
\alias{download_button}
\title{Create a download button or link}
\usage{
download_button(
  id,
  label = "Download",
  ...,
  class = NULL,
  icon = shiny::icon("download")
)
}
\arguments{
\item{id}{An input id.}

\item{label}{The label that should appear on the button.}

\item{...}{Other arguments to pass to the container tag function.}

\item{class}{Additional CSS classes to apply to the tag, if any.}

\item{icon}{An \code{\link[shiny:icon]{icon()}} to appear on the button. Default is \code{icon("download")}.}
}
\description{
Use these functions to create a download button or link; when clicked, it
will initiate a browser download. The filename and contents are specified by
the corresponding \code{\link[shiny:downloadHandler]{downloadHandler()}} defined in the server
function.
}
\section{Aliased from Shiny}{
 This function is an alias for \code{shiny::downloadButton()} and is included to maintain more consistent naming conventions in Shiny apps that use \pkg{bslib}. The documentation on this page may still refer to the original function names. You can replace \code{shiny::downloadButton()} with \code{download_button()}.
}

\seealso{
Other Shiny upload/download aliases: 
\code{\link{download_handler}()},
\code{\link{download_link}()},
\code{\link{input_file}()}
}
\concept{Shiny upload/download aliases}
