% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-download_button.R
\name{download_handler}
\alias{download_handler}
\title{File Downloads}
\usage{
download_handler(filename, content, contentType = NULL, outputArgs = list())
}
\arguments{
\item{filename}{A string of the filename, including extension, that the
user's web browser should default to when downloading the file; or a
function that returns such a string. (Reactive values and functions may be
used from this function.)}

\item{content}{A function that takes a single argument \code{file} that is a
file path (string) of a nonexistent temp file, and writes the content to
that file path. (Reactive values and functions may be used from this
function.)}

\item{contentType}{A string of the download's
\href{https://en.wikipedia.org/wiki/Internet_media_type}{content type}, for
example \code{"text/csv"} or \code{"image/png"}. If \code{NULL}, the content type
will be guessed based on the filename extension, or
\code{application/octet-stream} if the extension is unknown.}

\item{outputArgs}{A list of arguments to be passed through to the implicit
call to \code{\link[shiny:downloadButton]{downloadButton()}} when \code{downloadHandler} is used
in an interactive R Markdown document.}
}
\description{
Allows content from the Shiny application to be made available to the user as
file downloads (for example, downloading the currently visible data as a CSV
file). Both filename and contents can be calculated dynamically at the time
the user initiates the download. Assign the return value to a slot on
\code{output} in your server function, and in the UI use
\code{\link[shiny:downloadButton]{downloadButton()}} or \code{\link[shiny:downloadLink]{downloadLink()}} to make the
download available.
}
\section{Aliased from Shiny}{
 This function is an alias for \code{shiny::downloadHandler()} and is included to maintain more consistent naming conventions in Shiny apps that use \pkg{bslib}. The documentation on this page may still refer to the original function names. You can replace \code{shiny::downloadHandler()} with \code{download_handler()}.
}

\seealso{
\itemize{
\item \code{\link[=download_button]{download_button()}} creates a download button in the UI.
\item \code{\link[=download_link]{download_link()}} creates a download link in the UI.
}

Other Shiny upload/download aliases: 
\code{\link{download_button}()},
\code{\link{download_link}()},
\code{\link{input_file}()}
}
\concept{Shiny upload/download aliases}
