% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-download_button.R
\name{download_link}
\alias{download_link}
\title{Create a download button or link}
\usage{
download_link(id, label = "Download", ..., class = NULL)
}
\arguments{
\item{id}{An input id.}

\item{label}{The label that should appear on the button.}

\item{...}{Other arguments to pass to the container tag function.}

\item{class}{Additional CSS classes to apply to the tag, if any.}
}
\description{
Use these functions to create a download button or link; when clicked, it
will initiate a browser download. The filename and contents are specified by
the corresponding \code{\link[shiny:downloadHandler]{downloadHandler()}} defined in the server
function.
}
\section{Aliased from Shiny}{
 This function is an alias for \code{shiny::downloadLink()} and is included to maintain more consistent naming conventions in Shiny apps that use \pkg{bslib}. The documentation on this page may still refer to the original function names. You can replace \code{shiny::downloadLink()} with \code{download_link()}.
}

\seealso{
Other Shiny upload/download aliases: 
\code{\link{download_button}()},
\code{\link{download_handler}()},
\code{\link{input_file}()}
}
\concept{Shiny upload/download aliases}
