% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfonts.R
\name{gfont}
\alias{gfont}
\title{Import Google Fonts into a Bootstrap theme}
\usage{
gfont(name, weight = NULL, local = FALSE, cache = TRUE)
}
\arguments{
\item{name}{Any Google Font name (e.g., Pacifico)}

\item{weight}{A font weight definition}

\item{local}{whether or not download and bundle local (woff) font files.}

\item{cache}{whether or not to cache local font files (only takes effect
when \code{local = TRUE}).}
}
\description{
Import Google Fonts into a Bootstrap theme
}
\examples{

theme <- bs_theme(base_font = gfont("Pacifico"))
if (interactive()) {
  bs_theme_preview(theme)
}
}
\references{
\url{https://developers.google.com/fonts/docs/css2}
}
