% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.bslib_fragment}
\alias{print.bslib_fragment}
\alias{print.bslib_page}
\title{Make HTML browsable by default}
\usage{
\method{print}{bslib_fragment}(x, ...)

\method{print}{bslib_page}(x, ...)
}
\arguments{
\item{x}{a \code{\link[=tag]{tag()}} object.}

\item{...}{passed along to an underlying print method}
}
\description{
Make HTML browsable by default
}
\keyword{internal}
