% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input_action_button.R
\name{input_action_button}
\alias{input_action_button}
\title{Action button/link}
\usage{
input_action_button(
  id,
  label,
  ...,
  icon = NULL,
  width = NULL,
  disabled = FALSE
)
}
\arguments{
\item{id}{An input id.}

\item{label}{An input label.}

\item{...}{Named attributes to be applied to the button or link.}

\item{icon}{An optional \code{\link[shiny:icon]{icon()}} to appear on the button.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}

\item{disabled}{If \code{TRUE}, the button will not be clickable. Use
\code{\link[shiny:updateActionButton]{updateActionButton()}} to dynamically enable/disable the button.}
}
\description{
Creates an action button or link whose value is initially zero, and increments by one
each time it is pressed.
}
\section{Server value}{


An integer of class \code{"shinyActionButtonValue"}. This class differs from
ordinary integers in that a value of 0 is considered "falsy".
This implies two things:
\itemize{
\item Event handlers (e.g., \code{\link[shiny:observeEvent]{observeEvent()}}, \code{\link[shiny:eventReactive]{eventReactive()}}) won't execute on initial load.
\item Input validation (e.g., \code{\link[shiny:req]{req()}}, \code{\link[shiny:need]{need()}}) will fail on initial load.
}

}

\seealso{
Other Shiny input aliases: 
\code{\link{input_action_link}()},
\code{\link{input_select}()}
}
\concept{Shiny input aliases}
