% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input_action_link.R
\name{input_action_link}
\alias{input_action_link}
\title{Action button/link}
\usage{
input_action_link(id, label, icon = NULL, ...)
}
\arguments{
\item{id}{An input id.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{icon}{An optional \code{\link[shiny:icon]{icon()}} to appear on the button.}

\item{...}{Named attributes to be applied to the button or link.}
}
\description{
Creates an action button or link whose value is initially zero, and increments by one
each time it is pressed.
}
\section{Server value}{


An integer of class \code{"shinyActionButtonValue"}. This class differs from
ordinary integers in that a value of 0 is considered "falsy".
This implies two things:
\itemize{
\item Event handlers (e.g., \code{\link[shiny:observeEvent]{observeEvent()}}, \code{\link[shiny:eventReactive]{eventReactive()}}) won't execute on initial load.
\item Input validation (e.g., \code{\link[shiny:req]{req()}}, \code{\link[shiny:need]{need()}}) will fail on initial load.
}

}

\seealso{
Other Shiny input aliases: 
\code{\link{input_action_button}()},
\code{\link{input_date}()},
\code{\link{input_date_range}()},
\code{\link{input_select}()},
\code{\link{input_selectize}()},
\code{\link{input_slider}()}
}
\concept{Shiny input aliases}
