% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-button-group.R
\name{input_check_buttons}
\alias{input_check_buttons}
\title{Create a button group of radio/check boxes}
\usage{
input_check_buttons(
  id,
  choices,
  ...,
  selected = NULL,
  size = c("md", "sm", "lg"),
  bg = "primary",
  outline = TRUE
)
}
\description{
Use \code{input_check_buttons()} if multiple choices may be selected at once; otherwise, use \code{input_radio_buttons()}
}
