% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-button-group.R
\name{input_check_search}
\alias{input_check_search}
\title{Search a "large" number of check options}
\usage{
input_check_search(
  id,
  choices,
  selected = NULL,
  placeholder = "ğŸ” Search",
  height = "150px"
)
}
\description{
Search a "large" number of check options
}
