% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input_checkbox.R
\name{input_checkbox}
\alias{input_checkbox}
\title{Checkbox Input Control}
\usage{
input_checkbox(id, label, value = FALSE, ..., width = NULL)
}
\arguments{
\item{id}{An input id.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}

\item{...}{Ignored, included for future expansion.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}
}
\value{
A checkbox control that can be added to a UI definition.
}
\description{
Create a checkbox that can be used to specify logical values.
}
\section{Server value}{


\code{TRUE} if checked, \code{FALSE} otherwise.

}

\seealso{
Other Shiny input aliases: 
\code{\link{input_action_button}()},
\code{\link{input_action_link}()},
\code{\link{input_date}()},
\code{\link{input_date_range}()},
\code{\link{input_select}()},
\code{\link{input_selectize}()},
\code{\link{input_slider}()}
}
\concept{Shiny input aliases}
