% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input_numeric.R
\name{input_numeric}
\alias{input_numeric}
\title{Create a numeric input control}
\usage{
input_numeric(
  id,
  label,
  value,
  ...,
  min = NA,
  max = NA,
  step = NA,
  width = NULL
)
}
\arguments{
\item{id}{An input id.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{...}{Ignored, included for future expansion.}

\item{min}{Minimum allowed value}

\item{max}{Maximum allowed value}

\item{step}{Interval to use when stepping between min and max}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}
}
\value{
A numeric input control that can be added to a UI definition.
}
\description{
Create an input control for entry of numeric values
}
\section{Aliased from Shiny}{
 This function is an alias for \code{shiny::numericInput()} and is included to maintain more consistent naming conventions in Shiny apps that use \pkg{bslib}. The documentation may still refer to the original function names. You can replace \code{shiny::numericInput()} with \code{input_numeric()}.
}

\section{Server value}{


A numeric vector of length 1.

}

\seealso{
Other Shiny input aliases: 
\code{\link{input_action_button}()},
\code{\link{input_action_link}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_date}()},
\code{\link{input_date_range}()},
\code{\link{input_password}()},
\code{\link{input_radio_buttons}()},
\code{\link{input_select}()},
\code{\link{input_selectize}()},
\code{\link{input_slider}()},
\code{\link{input_text}()},
\code{\link{input_text_area}()}
}
\concept{Shiny input aliases}
