% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input_password.R
\name{input_password}
\alias{input_password}
\title{Create a password input control}
\usage{
input_password(id, label, value = "", ..., width = NULL, placeholder = NULL)
}
\arguments{
\item{id}{An input id.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{...}{Ignored, included for future expansion.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}
}
\value{
A text input control that can be added to a UI definition.
}
\description{
Create an password control for entry of passwords.
}
\section{Server value}{


A character string of the password input. The default value is \code{""}
unless \code{value} is provided.

}

\seealso{
Other Shiny input aliases: 
\code{\link{input_action_button}()},
\code{\link{input_action_link}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_date}()},
\code{\link{input_date_range}()},
\code{\link{input_numeric}()},
\code{\link{input_radio_buttons}()},
\code{\link{input_select}()},
\code{\link{input_selectize}()},
\code{\link{input_slider}()},
\code{\link{input_text}()},
\code{\link{input_text_area}()}
}
\concept{Shiny input aliases}
