% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-submit.R
\name{input_submit_button}
\alias{input_submit_button}
\title{Suspend input changes until a button is clicked}
\usage{
input_submit_button(id, label, ..., scope = NULL)
}
\arguments{
\item{id}{The input ID.}

\item{label}{A label to place on the button.}

\item{...}{Arguments passed along to \code{\link[=input_task_button]{input_task_button()}}.}

\item{scope}{The scope of the submit button. Can be one of the following:
\itemize{
\item \code{NULL}: Inputs that are children of the button's parent are
deferred until the button is clicked.
\item A CSS selector: Only inputs that are within the element matching the
selector are deferred until the button is clicked.
}}
}
\description{
Suspend changes to a particular set of input controls until a submit button
is clicked. This is particularly useful for allowing the user to review their
input(s) before sending them to the server for a potentially expensive
operation. Note that, by default, all inputs that are children of the
button's parent are deferred until the button is clicked. This can be changed
by setting the \code{scope} argument to a CSS selector that matches the container
of inputs you wish to suspend.
}
\seealso{
\code{\link[=input_submit_textarea]{input_submit_textarea()}}, \code{\link[=input_task_button]{input_task_button()}}
}
