% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input_text.R
\name{input_text}
\alias{input_text}
\title{Create a text input control}
\usage{
input_text(id, label, value = "", ..., width = NULL, placeholder = NULL)
}
\arguments{
\item{id}{An input id.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{...}{Ignored, included for future expansion.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}
}
\value{
A text input control that can be added to a UI definition.
}
\description{
Create an input control for entry of unstructured text values
}
\section{Aliased from Shiny}{
 This function is an alias for \code{shiny::textInput()} and is included to maintain more consistent naming conventions in Shiny apps that use \pkg{bslib}. The documentation on this page may still refer to the original function names. You can replace \code{shiny::textInput()} with \code{input_text()}.
}

\section{Server value}{


A character string of the text input. The default value is \code{""}
unless \code{value} is provided.

}

\seealso{
\code{\link[=update_text]{update_text()}} to programmatically update a text input.

Other Shiny input aliases: 
\code{\link{input_action_button}()},
\code{\link{input_action_link}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_date}()},
\code{\link{input_date_range}()},
\code{\link{input_numeric}()},
\code{\link{input_password}()},
\code{\link{input_radio_buttons}()},
\code{\link{input_select}()},
\code{\link{input_selectize}()},
\code{\link{input_slider}()},
\code{\link{input_text_area}()}
}
\concept{Shiny input aliases}
