% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_columns}
\alias{layout_columns}
\title{Bootstrap grid layout}
\usage{
layout_columns(
  ...,
  col_widths = NA,
  row_heights = 1,
  gap = "1rem",
  fill = TRUE,
  fillable = TRUE,
  class = NULL,
  height = NULL,
  width = NULL
)
}
\description{
TODO: describe me
}
