% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav-update.R
\name{nav_select}
\alias{nav_select}
\alias{nav_insert}
\alias{nav_prepend}
\alias{nav_append}
\alias{nav_remove}
\alias{nav_show}
\alias{nav_hide}
\title{Programmatically update panels}
\usage{
nav_select(id, selected = NULL, session = getDefaultReactiveDomain())

nav_insert(
  id,
  nav,
  target,
  position = c("before", "after"),
  select = FALSE,
  session = getDefaultReactiveDomain()
)

nav_prepend(
  id,
  nav,
  select = FALSE,
  menuName = NULL,
  session = getDefaultReactiveDomain()
)

nav_append(
  id,
  nav,
  select = FALSE,
  menuName = NULL,
  session = getDefaultReactiveDomain()
)

nav_remove(id, target, session = getDefaultReactiveDomain())

nav_show(id, target, select = FALSE, session = getDefaultReactiveDomain())

nav_hide(id, target, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{a character string identifying the nav container (e.g., \code{\link[=navs_tab]{navs_tab()}}).}

\item{selected}{a character string matching a particular \code{\link[=nav]{nav()}} of interest.}

\item{session}{a shiny session object (the default should almost always be used).}

\item{nav}{a \code{\link[=nav]{nav()}} item to insert.}

\item{target}{If inserting: the \code{value} of an existing \code{nav()} item, next to which tab will be added. If removing: the \code{value} of the \code{nav()} item that you want to remove.}

\item{position}{Should \code{nav} be added before or after the target?}

\item{select}{Should \code{nav} be selected upon being inserted?}
}
\description{
Functions for dynamically updating selected (i.e., active) \code{\link[=nav]{nav()}} items
within a particular nav container. These functions require an container \code{id}
to identify the container interest.
}
