% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navs.R
\name{nav}
\alias{nav}
\alias{navs_tab_card}
\alias{navs_bar}
\title{Create navigation items}
\usage{
nav(title, ..., value = title, icon = NULL)

navs_tab_card(..., id = NULL, selected = NULL, header = NULL, footer = NULL)

navs_bar(
  ...,
  title = NULL,
  id = NULL,
  selected = NULL,
  position = c("static-top", "fixed-top", "fixed-bottom"),
  header = NULL,
  footer = NULL,
  collapsible = TRUE,
  bg = NULL
)
}
\arguments{
\item{title}{\verb{[character(1)]}\cr A display title for the nav item.}

\item{...}{\verb{[UI elements]}\cr
\itemize{
\item For \code{tabPanel()}/\code{tabPanelBody()}: content to display when the tab is
active.
\item For \code{navItem()}: UI elements to place inside the nav (e.g., hyperlinks
\code{\link[shiny:reexports]{a()}}, \code{\link[shiny:actionButton]{actionButton()}}s, \code{\link[shiny:textInput]{textInput()}}, etc).
}}

\item{value}{\verb{[character(1)]}\cr A value associate with a \code{tabPanel()} to
determine when it is selected. This value may be supplied to the relevant
nav container's \code{selected} argument to show the tab immediately on page
load. This value is also useful for programmatically updating the selected
tab via \code{\link[shiny:updateTabsetPanel]{updateTabsetPanel()}}, \code{\link[shiny:showTab]{showTab()}}, etc (updating selected tabs
this way is often useful for showing/hiding panels of content via other UI
controls like \code{\link[shiny:radioButtons]{radioButtons()}} -- in this scenario, consider using
\code{\link[shiny:tabPanel]{tabPanelBody()}} with \code{tabsetPanel(type = "hidden")}).}

\item{icon}{An \code{\link[shiny:icon]{icon()}} to appear to the left of the nav item's \code{title}.}

\item{id}{If provided, you can use \verb{input$}\emph{\code{id}} in your
server logic to determine which of the current tabs is active. The value
will correspond to the \code{value} argument that is passed to
\code{\link[shiny:tabPanel]{tabPanel()}}.}

\item{selected}{The \code{value} (or, if none was supplied, the \code{title})
of the tab that should be selected by default. If \code{NULL}, the first
tab will be selected.}

\item{header}{Tag or list of tags to display as a common header above all
tabPanels.}

\item{footer}{Tag or list of tags to display as a common footer below all
tabPanels}

\item{position}{Determines whether the navbar should be displayed at the top
of the page with normal scrolling behavior (\code{"static-top"}), pinned at
the top (\code{"fixed-top"}), or pinned at the bottom
(\code{"fixed-bottom"}). Note that using \code{"fixed-top"} or
\code{"fixed-bottom"} will cause the navbar to overlay your body content,
unless you add padding, e.g.: \code{tags$style(type="text/css", "body
  {padding-top: 70px;}")}}

\item{collapsible}{\code{TRUE} to automatically collapse the navigation
elements into a menu when the width of the browser is less than 940 pixels
(useful for viewing on smaller touchscreen device)}
}
\description{
Create navigation items
}
\seealso{
\link{navs_tab}
}
