% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav-update.R
\name{nav_update}
\alias{nav_update}
\alias{nav_insert}
\alias{nav_remove}
\title{Programmatic panel updating}
\usage{
nav_update(
  input_id,
  selected = NULL,
  session = shiny::getDefaultReactiveDomain()
)

nav_insert(
  input_id,
  nav_item,
  target,
  position = c("before", "after"),
  select = FALSE,
  session = shiny::getDefaultReactiveDomain()
)

nav_remove(input_id, target, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{selected}{The \code{value} (or, if none was supplied, the \code{title})
of the tab that should be selected by default. If \code{NULL}, the first
tab will be selected.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}

\item{target}{If inserting: the \code{value} of an existing
\code{tabPanel}, next to which \code{tab} will be added.
If removing: the \code{value} of the \code{tabPanel} that
you want to remove. See Details if you want to insert next to/remove
an entire \code{navbarMenu} instead.}

\item{position}{Should \code{tab} be added before or after the
\code{target} tab?}

\item{select}{Should \code{tab} be selected upon being inserted?}
}
\description{
Programmatic panel updating
}
