% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{navbar_height}
\alias{navbar_height}
\title{Height of navbar height for a given Bootswatch theme}
\usage{
navbar_height(theme = "", version = version_latest())
}
\arguments{
\item{theme}{a \url{https://bootswatch.com/} theme name. The default value
means vanilla Bootstrap (i.e. no bootswatch theme).}
}
\description{
The height (in pixels) of a Bootswatch theme's navbar. Note that these
heights to do not currently respect theme customizations.
}
\details{
This is primarily useful for logic internal to \code{\link[rmarkdown:html_document]{rmarkdown::html_document()}}
}
