% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navs.R
\name{navs_tab}
\alias{navs_tab}
\alias{navs_pill}
\alias{navs_pill_card}
\alias{navs_pill_list}
\alias{navs_hidden}
\title{Create a navigation panel}
\usage{
navs_tab(..., id = NULL, selected = NULL, header = NULL, footer = NULL)

navs_pill(..., id = NULL, selected = NULL, header = NULL, footer = NULL)

navs_pill_card(..., id = NULL, selected = NULL, header = NULL, footer = NULL)

navs_pill_list(..., id = NULL, selected = NULL, header = NULL, footer = NULL)

navs_hidden(..., id = NULL, selected = NULL)
}
\arguments{
\item{...}{\code{\link[shiny:tabPanel]{tabPanel()}} elements to include in the tabset}

\item{id}{If provided, you can use \verb{input$}\emph{\code{id}} in your
server logic to determine which of the current tabs is active. The value
will correspond to the \code{value} argument that is passed to
\code{\link[shiny:tabPanel]{tabPanel()}}.}

\item{selected}{The \code{value} (or, if none was supplied, the \code{title})
of the tab that should be selected by default. If \code{NULL}, the first
tab will be selected.}

\item{header}{Tag or list of tags to display as a common header above all
tabPanels.}

\item{footer}{Tag or list of tags to display as a common footer below all
tabPanels}
}
\description{
Render a collection of \code{\link[=nav]{nav()}} items into a panel of contents.
}
