% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-output_table.R
\name{output_table}
\alias{output_table}
\title{Table Output}
\usage{
output_table(id)
}
\description{
The \code{tableOuptut()}/\code{renderTable()} pair creates a reactive table that is
suitable for display small matrices and data frames. The columns are
formatted with \code{\link[xtable:xtable]{xtable::xtable()}}.

See \code{\link[shiny:renderDataTable]{renderDataTable()}} for data frames that are too big to fit on a single
page.
}
\seealso{
\code{\link[=render_table]{render_table()}} to reactively update the \code{new_output()}.

Other Shiny output aliases: 
\code{\link{output_image}()},
\code{\link{output_plot}()},
\code{\link{output_text}()},
\code{\link{output_text_verbatim}()},
\code{\link{output_ui}()}
}
\concept{Shiny output aliases}
