% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-output_text.R
\name{output_text}
\alias{output_text}
\title{Create a text output element}
\usage{
output_text(id, inline = FALSE, container = if (inline) span else div)
}
\arguments{
\item{id}{An output id.}

\item{inline}{use an inline (\code{span()}) or block container (\code{div()})
for the output}

\item{container}{a function to generate an HTML element to contain the text}
}
\value{
An output element for use in UI.
}
\description{
Render a reactive output variable as text within an application page.
\code{textOutput()} is usually paired with \code{\link[shiny:renderText]{renderText()}} and puts regular text
in \verb{<div>} or \verb{<span>}; \code{verbatimTextOutput()} is usually paired with
\code{\link[shiny:renderPrint]{renderPrint()}} and provides fixed-width text in a \verb{<pre>}.
}
\details{
In both functions, text is HTML-escaped prior to rendering.
}
\seealso{
\code{\link[=render_text]{render_text()}} to reactively update the \code{new_output()}.

Other Shiny output aliases: 
\code{\link{output_image}()},
\code{\link{output_plot}()},
\code{\link{output_table}()},
\code{\link{output_text_verbatim}()},
\code{\link{output_ui}()}
}
\concept{Shiny output aliases}
