% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-output_text_verbatim.R
\name{output_text_verbatim}
\alias{output_text_verbatim}
\title{Create a text output element}
\usage{
output_text_verbatim(id, placeholder = FALSE)
}
\arguments{
\item{placeholder}{if the output is empty or \code{NULL}, should an empty
rectangle be displayed to serve as a placeholder? (does not affect
behavior when the output is nonempty)}
}
\value{
An output element for use in UI.
}
\description{
Render a reactive output variable as text within an application page.
\code{textOutput()} is usually paired with \code{\link[shiny:renderText]{renderText()}} and puts regular text
in \verb{<div>} or \verb{<span>}; \code{verbatimTextOutput()} is usually paired with
\code{\link[shiny:renderPrint]{renderPrint()}} and provides fixed-width text in a \verb{<pre>}.
}
\details{
In both functions, text is HTML-escaped prior to rendering.
}
\seealso{
\code{\link[=render_text_verbatim]{render_text_verbatim()}} to reactively update the \code{new_output()}.

Other Shiny output aliases: 
\code{\link{output_image}()},
\code{\link{output_plot}()},
\code{\link{output_table}()},
\code{\link{output_text}()},
\code{\link{output_ui}()}
}
\concept{Shiny output aliases}
