% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-output_ui.R
\name{output_ui}
\alias{output_ui}
\title{Create an HTML output element}
\usage{
output_ui(
  id,
  ...,
  inline = FALSE,
  container = if (inline) span else div,
  fill = FALSE
)
}
\arguments{
\item{id}{An output id.}

\item{...}{Other arguments to pass to the container tag function. This is
useful for providing additional classes for the tag.}

\item{inline}{use an inline (\code{span()}) or block container (\code{div()})
for the output}

\item{container}{a function to generate an HTML element to contain the text}

\item{fill}{If \code{TRUE}, the result of \code{container} is treated as \emph{both} a fill
item and container (see \code{\link[htmltools:bindFillRole]{htmltools::bindFillRole()}}), which means both the
\code{container} as well as its immediate children (i.e., the result of
\code{renderUI()}) are allowed to grow/shrink to fit a fill container with an
opinionated height. Set \code{fill = "item"} or \code{fill = "container"} to treat
\code{container} as just a fill item or a fill container.}
}
\value{
An HTML output element that can be included in a panel
}
\description{
Render a reactive output variable as HTML within an application page. The
text will be included within an HTML \code{div} tag, and is presumed to contain
HTML content which should not be escaped.
}
\details{
\code{uiOutput} is intended to be used with \code{renderUI} on the server side. It is
currently just an alias for \code{htmlOutput}.
}
\seealso{
\code{\link[=render_ui]{render_ui()}} to reactively update the \code{new_output()}.

Other Shiny output aliases: 
\code{\link{output_image}()},
\code{\link{output_plot}()},
\code{\link{output_table}()},
\code{\link{output_text}()},
\code{\link{output_text_verbatim}()}
}
\concept{Shiny output aliases}
