% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page_fixed}
\alias{page_fixed}
\title{Create a page with fluid layout}
\usage{
page_fixed(..., title = NULL, theme = bs_theme(), lang = NULL)
}
\arguments{
\item{...}{Elements to include within the container}

\item{title}{The browser window title (defaults to the host URL of the page)}

\item{theme}{One of the following:
\itemize{
\item \code{NULL} (the default), which implies a "stock" build of Bootstrap 3.
\item A \code{\link[bslib:bs_theme]{bslib::bs_theme()}} object. This can be used to replace a stock
build of Bootstrap 3 with a customized version of Bootstrap 3 or higher.
\item A character string pointing to an alternative Bootstrap stylesheet
(normally a css file within the www directory, e.g. \code{www/bootstrap.css}).
}}

\item{lang}{ISO 639-1 language code for the HTML page, such as "en" or "ko".
This will be used as the lang in the \code{<html>} tag, as in \code{<html lang="en">}.
The default (NULL) results in an empty string.}
}
\description{
Alias for \code{\link[shiny:fixedPage]{shiny::fixedPage()}} with \code{theme} defaulting to bslib's recommended
version Bootstrap.
}
