% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page_fluid}
\alias{page_fluid}
\title{Create a page with fluid layout}
\usage{
page_fluid(..., title = NULL, theme = bs_theme(), lang = NULL)
}
\arguments{
\item{...}{Elements to include within the page}

\item{title}{The browser window title (defaults to the host URL of the page).
Can also be set as a side effect of the \code{\link[shiny:titlePanel]{titlePanel()}} function.}

\item{theme}{Alternative Bootstrap stylesheet (normally a css file within the
www directory). For example, to use the theme located at
\code{www/bootstrap.css} you would use \code{theme = "bootstrap.css"}.}

\item{lang}{ISO 639-1 language code for the HTML page, such as "en" or "ko".
This will be used as the lang in the \code{<html>} tag, as in \code{<html lang="en">}.
The default (NULL) results in an empty string.}
}
\description{
Alias for \code{\link[shiny:fluidPage]{shiny::fluidPage()}} with \code{theme} defaulting to bslib's recommended
version Bootstrap.
}
