% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page_sidebar}
\alias{page_sidebar}
\title{A sidebar page (i.e., dashboard)}
\usage{
page_sidebar(
  ...,
  sidebar = sidebar(),
  title = NULL,
  fillable = TRUE,
  fill_mobile = FALSE,
  theme = bs_theme(),
  window_title = NA,
  lang = NULL
)
}
\arguments{
\item{...}{The contents of the document body.}

\item{sidebar}{A \code{\link[=sidebar]{sidebar()}} component to display on every \code{\link[=nav_panel]{nav_panel()}}
page.}

\item{title}{A string, number, or \code{\link[htmltools:builder]{htmltools::tag()}} child to display as the
title (just above the \code{\link[=layout_sidebar]{layout_sidebar()}}).}

\item{fillable}{Whether or not to allow \code{fill} items to grow/shrink to fit
the browser window. If \code{TRUE}, all \code{\link[=nav_panel]{nav_panel()}} pages are \code{fillable}. A
character vector, matching the \code{value} of \code{\link[=nav_panel]{nav_panel()}}s to be filled, may
also be provided. Note that, if a \code{sidebar} is provided, \code{fillable} makes
the main content portion fillable.}

\item{fill_mobile}{Whether or not \code{fillable} pages should fill the viewport's
height on mobile devices (i.e., narrow windows).}

\item{theme}{A \code{\link[=bs_theme]{bs_theme()}} object.}

\item{window_title}{the browser window title. The default value, \code{NA}, means
to use any character strings that appear in \code{title} (if none are found, the
host URL of the page is displayed by default).}

\item{lang}{ISO 639-1 language code for the HTML page, such as "en" or "ko".
This will be used as the lang in the \code{<html>} tag, as in \code{<html lang="en">}.
The default (NULL) results in an empty string.}
}
\description{
Easily create a traditional dashboard layout with a full-bleed header
(\code{title}) and \code{\link[=sidebar]{sidebar()}}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(shiny)
library(ggplot2)

ui <- page_sidebar(
  title = "Example dashboard",
  sidebar = sidebar(
    varSelectInput("var", "Select variable", mtcars)
  ),
  card(
    full_screen = TRUE,
    card_header("My plot"),
    plotOutput("p")
  )
)

server <- function(input, output) {
  output$p <- renderPlot({
    ggplot(mtcars) + geom_histogram(aes(!!input$var))
  })
}

shinyApp(ui, server)
\dontshow{\}) # examplesIf}
}
