% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebuilt.R
\name{prebuilt_css_path}
\alias{prebuilt_css_path}
\title{Get the path to a pre-built CSS file}
\usage{
prebuilt_css_path(theme)
}
\arguments{
\item{theme}{A theme object.}
}
\value{
The path to the pre-build CSS file, if it exists, or \code{NULL}, if not.
}
\description{
Get the path to a pre-built CSS file
}
\examples{
# The path to a pre-built CSS file for a theme
prebuilt_css_path(bs_theme_create("4"))

# Attempting to get the path to a theme that is not pre-built will return NULL
prebuilt_css_path(bs_theme_create(bootswatch = "sketchy"))
}
