% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-output_image.R
\name{render_image}
\alias{render_image}
\title{Image file output}
\usage{
render_image(
  expr,
  env = parent.frame(),
  quoted = FALSE,
  deleteFile,
  outputArgs = list()
)
}
\arguments{
\item{expr}{An expression that returns a list.}

\item{env}{The parent environment for the reactive expression. By default,
this is the calling environment, the same as when defining an ordinary
non-reactive expression. If \code{expr} is a quosure and \code{quoted} is \code{TRUE},
then \code{env} is ignored.}

\item{quoted}{If it is \code{TRUE}, then the \code{\link[=quote]{quote()}}ed value of \code{expr}
will be used when \code{expr} is evaluated. If \code{expr} is a quosure and you
would like to use its expression as a value for \code{expr}, then you must set
\code{quoted} to \code{TRUE}.}

\item{deleteFile}{Should the file in \code{func()$src} be deleted after
it is sent to the client browser? Generally speaking, if the image is a
temp file generated within \code{func}, then this should be \code{TRUE};
if the image is not a temp file, this should be \code{FALSE}. (For backward
compatibility reasons, if this argument is missing, a warning will be
emitted, and if the file is in the temp directory it will be deleted. In
the future, this warning will become an error.)}

\item{outputArgs}{A list of arguments to be passed through to the implicit
call to \code{\link[shiny:imageOutput]{imageOutput()}} when \code{renderImage} is used in an
interactive R Markdown document.}
}
\description{
Renders a reactive image that is suitable for assigning to an \code{output}
slot.
}
\details{
The expression \code{expr} must return a list containing the attributes for
the \code{img} object on the client web page. For the image to display,
properly, the list must have at least one entry, \code{src}, which is the
path to the image file. It may also useful to have a \code{contentType}
entry specifying the MIME type of the image. If one is not provided,
\code{renderImage} will try to autodetect the type, based on the file
extension.

Other elements such as \code{width}, \code{height}, \code{class}, and
\code{alt}, can also be added to the list, and they will be used as
attributes in the \code{img} object.

The corresponding HTML output tag should be \code{div} or \code{img} and have
the CSS class name \code{shiny-image-output}.
}
\section{Aliased from Shiny}{
 This function is an alias for \code{shiny::renderImage()} and is included to maintain more consistent naming conventions in Shiny apps that use \pkg{bslib}. The documentation on this page may still refer to the original function names. You can replace \code{shiny::renderImage()} with \code{render_image()}.
}

\seealso{
\code{\link[=output_image]{output_image()}} to create an output in the UI.

Other Shiny render aliases: 
\code{\link{render_plot}()},
\code{\link{render_table}()},
\code{\link{render_text}()},
\code{\link{render_text_verbatim}()},
\code{\link{render_ui}()}
}
\concept{Shiny render aliases}
