% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-output_text_verbatim.R
\name{render_text_verbatim}
\alias{render_text_verbatim}
\title{Text Output}
\usage{
render_text_verbatim(
  expr,
  env = parent.frame(),
  quoted = FALSE,
  width = getOption("width"),
  outputArgs = list()
)
}
\arguments{
\item{expr}{An expression to evaluate.}

\item{env}{The parent environment for the reactive expression. By default,
this is the calling environment, the same as when defining an ordinary
non-reactive expression. If \code{expr} is a quosure and \code{quoted} is \code{TRUE},
then \code{env} is ignored.}

\item{quoted}{If it is \code{TRUE}, then the \code{\link[=quote]{quote()}}ed value of \code{expr}
will be used when \code{expr} is evaluated. If \code{expr} is a quosure and you
would like to use its expression as a value for \code{expr}, then you must set
\code{quoted} to \code{TRUE}.}

\item{width}{Width of printed output.}

\item{outputArgs}{A list of arguments to be passed through to the implicit
call to \code{\link[shiny:verbatimTextOutput]{verbatimTextOutput()}} or \code{\link[shiny:textOutput]{textOutput()}} when the functions are
used in an interactive RMarkdown document.}
}
\value{
For \code{renderPrint()}, note the given expression returns \code{NULL} then \code{NULL}
will actually be visible in the output. To display nothing, make your
function return \code{\link[=invisible]{invisible()}}.
}
\description{
\code{renderPrint()} prints the result of \code{expr}, while \code{renderText()} pastes it
together into a single string. \code{renderPrint()} is equivalent to \code{\link[=print]{print()}};
\code{renderText()} is equivalent to \code{\link[=cat]{cat()}}. Both functions capture all other
printed output generated while evaluating \code{expr}.

\code{renderPrint()} is usually paired with \code{\link[shiny:verbatimTextOutput]{verbatimTextOutput()}};
\code{renderText()} is usually paired with \code{\link[shiny:textOutput]{textOutput()}}.
}
\details{
The corresponding HTML output tag can be anything (though \code{pre} is
recommended if you need a monospace font and whitespace preserved) and should
have the CSS class name \code{shiny-text-output}.
}
\section{Aliased from Shiny}{
 This function is an alias for \code{shiny::renderPrint()} and is included to maintain more consistent naming conventions in Shiny apps that use \pkg{bslib}. The documentation may still refer to the original function names. You can replace \code{shiny::renderPrint()} with \code{render_text_verbatim()}.
}

\seealso{
\code{\link[=output_text_verbatim]{output_text_verbatim()}} to create an output in the UI.

Other Shiny render aliases: 
\code{\link{render_image}()},
\code{\link{render_plot}()},
\code{\link{render_table}()},
\code{\link{render_text}()},
\code{\link{render_ui}()}
}
\concept{Shiny render aliases}
