% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-output_ui.R
\name{render_ui}
\alias{render_ui}
\title{UI Output}
\usage{
render_ui(expr, env = parent.frame(), quoted = FALSE, outputArgs = list())
}
\arguments{
\item{expr}{An expression that returns a Shiny tag object, \code{\link[shiny:HTML]{HTML()}},
or a list of such objects.}

\item{env}{The parent environment for the reactive expression. By default,
this is the calling environment, the same as when defining an ordinary
non-reactive expression. If \code{expr} is a quosure and \code{quoted} is \code{TRUE},
then \code{env} is ignored.}

\item{quoted}{If it is \code{TRUE}, then the \code{\link[=quote]{quote()}}ed value of \code{expr}
will be used when \code{expr} is evaluated. If \code{expr} is a quosure and you
would like to use its expression as a value for \code{expr}, then you must set
\code{quoted} to \code{TRUE}.}

\item{outputArgs}{A list of arguments to be passed through to the implicit
call to \code{\link[shiny:uiOutput]{uiOutput()}} when \code{renderUI} is used in an
interactive R Markdown document.}
}
\description{
Renders reactive HTML using the Shiny UI library.
}
\details{
The corresponding HTML output tag should be \code{div} and have the CSS class
name \code{shiny-html-output} (or use \code{\link[shiny:uiOutput]{uiOutput()}}).
}
\section{Aliased from Shiny}{
 This function is an alias for \code{shiny::renderUI()} and is included to maintain more consistent naming conventions in Shiny apps that use \pkg{bslib}. The documentation may still refer to the original function names. You can replace \code{shiny::renderUI()} with \code{render_ui()}.
}

\seealso{
\code{\link[=output_ui]{output_ui()}} to create an output in the UI.

Other Shiny render aliases: 
\code{\link{render_image}()},
\code{\link{render_plot}()},
\code{\link{render_table}()},
\code{\link{render_text}()},
\code{\link{render_text_verbatim}()}
}
\concept{Shiny render aliases}
