% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_sass.R
\name{sass_file_bootstrap}
\alias{sass_file_bootstrap}
\alias{sass_file_bootswatch}
\title{Obtain Bootstrap and Bootswatch SASS files}
\usage{
sass_file_bootstrap(file = NULL, version = 4)

sass_file_bootswatch(theme, file = NULL, version = version_latest())
}
\arguments{
\item{file}{a scss file path.}

\item{version}{the major version.}

\item{theme}{a bootswatch theme name.}
}
\description{
Useful if you need to import specific SASS files from
Bootstrap and/or Bootswatch as part of a \code{\link[=bs_sass_partial]{bs_sass_partial()}}
based workflow.
}
\details{
\code{sass_file_bootstrap()} defaults to the main Bootstrap scss file.
\code{sass_file_bootswatch()} default to the variables scss file.
}
