% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sass_variables.R
\name{sass_variables_add}
\alias{sass_variables_add}
\alias{sass_variables_clear}
\alias{sass_variables}
\title{Sets SASS variables globally}
\usage{
sass_variables_add(...)

sass_variables_clear()

sass_variables()
}
\arguments{
\item{...}{Variables to be set, using \code{name = value}; or, a single unnamed
argument which is a named list of variables to set. To specify a name that
uses characters that aren't valid for R identifiers, wrap that name in
quotes, double quotes, or backticks (see examples section below).}
}
\value{
If no arguments are passed, a list of all set options is returned. If
arguments are passed, a list containing the previous values of the newly
set options is returned invisibly.
}
\description{
Set Sass variables that will be respected by \code{\link[=bs4_sass]{bs4_sass()}} (by default).
}
\examples{
library(htmltools)

sass_variables_add(primary = "orange", "body-bg" = "#EEEEEE")
sass_variables_add("font-family-base" = "monospace", "font-size-base" = "1.4rem")
sass_variables_add("btn-padding-y" = ".16rem")
sass_variables_add("btn-padding-x" = "2rem")
sass_variables_add("border-radius" = 0, "border-radius-lg" = 0, "border-radius-sm" = 0)

browsable(tags$body(
  bs4_sass(bootswatch_theme = "cosmo"),
  tags$a(class = "btn btn-primary", href = "#", role = "button", "Hello")
))

}
