% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{save_html.bslib_fragment}
\alias{save_html.bslib_fragment}
\alias{save_html.bslib_page}
\title{Save a bslib page/fragment as HTML}
\usage{
\method{save_html}{bslib_fragment}(html, file, ...)

\method{save_html}{bslib_page}(html, file, ...)
}
\arguments{
\item{html}{a bslib page/fragment.}

\item{...}{passed along to an underlying \code{\link[htmltools:save_html]{htmltools::save_html()}} method.}
}
\description{
Save a bslib page/fragment as HTML
}
\keyword{internal}
