% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{show_toast}
\alias{show_toast}
\title{Display a toast notification}
\usage{
show_toast(toast, ..., session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{toast}{A \code{bslib_toast} object created by \code{\link[=toast]{toast()}}, or a string/UI
element (which will be automatically converted to a toast with default
settings).}

\item{...}{Reserved for future extensions (currently ignored).}

\item{session}{Shiny session object.}
}
\value{
Invisibly returns the toast ID (string) that can be used with
\code{\link[=hide_toast]{hide_toast()}}.
}
\description{
Displays a toast notification in a Shiny application.
}
\examples{
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}
library(shiny)
library(bslib)

ui <- page_fluid(
  actionButton("show_simple", "Show Simple Toast"),
  actionButton("show_success", "Show Success Toast")
)

server <- function(input, output, session) {
  observeEvent(input$show_simple, {
    # Simple string automatically converted to toast
    show_toast("Operation completed!")
  })

  observeEvent(input$show_success, {
    # Show a pre-created toast
    show_toast(
      toast(
        body = "Your file has been uploaded.",
        header = "Success",
        type = "success"
      )
    )
  })
}

shinyApp(ui, server)
\dontshow{\}) # examplesIf}
}
\seealso{
Other Toast components: 
\code{\link{hide_toast}()},
\code{\link{toast}()},
\code{\link{toast_header}()}
}
\concept{Toast components}
