% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value-box.R
\name{showcase_left_center}
\alias{showcase_left_center}
\alias{showcase_top_right}
\alias{showcase_bottom}
\title{Custom layouts to use in other functions}
\usage{
showcase_left_center(
  width = 0.3,
  width_full_screen = "1fr",
  max_height = "100px",
  max_height_full_screen = 0.67
)

showcase_top_right(
  width = 0.4,
  width_full_screen = "1fr",
  max_height = "75px",
  max_height_full_screen = 0.67
)

showcase_bottom(
  width = "100\%",
  width_full_screen = NULL,
  height = "auto",
  height_full_screen = "2fr",
  max_height = "100px",
  max_height_full_screen = NULL
)
}
\arguments{
\item{width, width_full_screen, height, height_full_screen}{one of the
following:
\itemize{
\item A proportion (i.e., a number between 0 and 1) of available width or
height to allocate to the showcase.
\item A valid \link[htmltools:validateCssUnit]{CSS unit} defining the width or
height of the showcase column, or a valid value accepted by the
\code{grid-template-columns} (width) or \code{grid-template-rows} (height) CSS
property to define the width or height of the showcase column or row.
Accepted values in the second category are \code{"auto"}, \code{"min-content"},
\code{"max-content"}, a fractional unit (e.g. \verb{2fr}), or a \code{minmax()} function
(e.g., \verb{minmax(100px, 1fr)}).
}}

\item{max_height_full_screen}{A proportion (i.e., a number between
0 and 1) or any valid \link[htmltools:validateCssUnit]{CSS unit} defining the
showcase \code{max_height} in a full screen card.}
}
\description{
An collection of functions to build custom layouts that can be used in other
functions, such as \code{\link[=value_box]{value_box()}}.
}
\examples{
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}
library(htmltools)

value_box(
  "KPI Title",
  h1(HTML("$1 <i>Billion</i> Dollars")),
  span(
    bsicons::bs_icon("arrow-up"),
    " 30\% VS PREVIOUS 30 DAYS"
  ),
  showcase = bsicons::bs_icon("piggy-bank"),
  showcase_layout = showcase_top_right()
)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=value_box]{value_box()}} takes a \code{showcase_layout} where one can use the various
\verb{showcase_*()} functions.

Other Components:
\code{\link[=accordion]{accordion()}},
\code{\link[=card]{card()}},
\code{\link[=popover]{popover()}},
\code{\link[=tooltip]{tooltip()}},
\code{\link[=value_box]{value_box()}}
}
\concept{Components}
