% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebar_toggle}
\alias{sidebar_toggle}
\title{Toggle the sidebar state (internal alias)}
\usage{
sidebar_toggle(id, open = NULL, session = get_current_session())
}
\arguments{
\item{id}{A character string. Required if wanting to re-actively read (or
update) the \code{collapsible} state in a Shiny app.}

\item{open}{The initial state of the sidebar, choosing from the following
options:
\itemize{
\item \code{"desktop"}: The sidebar starts open on desktop screen, closed on mobile.
This is default sidebar behavior.
\item \code{"open"} or \code{TRUE}: The sidebar starts open.
\item \code{"closed"} or \code{FALSE}: The sidebar starts closed.
\item \code{"always"} or \code{NA}: The sidebar is always open and cannot be closed.
}

In \code{sidebar_toggle()}, \code{open} indicates the desired state of the sidebar,
where the default of \code{open = NULL} will cause the sidebar to be toggled
open if closed or vice versa. Note that \code{sidebar_toggle()} can only open or
close the sidebar, so it does not support the \code{"desktop"} and \code{"always"}
options.}

\item{session}{A Shiny session object (the default should almost always be
used).}
}
\description{
Toggle the sidebar state (internal alias)
}
\keyword{internal}
