% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-theme-preset-bootswatch.R
\name{theme_bootswatch}
\alias{theme_bootswatch}
\title{Obtain a theme's Bootswatch theme name}
\usage{
theme_bootswatch(theme)
}
\arguments{
\item{theme}{A \code{\link[=bs_theme]{bs_theme()}} object.}
}
\value{
Returns the Bootswatch theme named used (if any) in the \code{theme}.
}
\description{
Obtain a theme's Bootswatch theme name
}
\seealso{
Other Bootstrap theme utility functions:
\code{\link[=bootswatch_themes]{bootswatch_themes()}},
\code{\link[=bs_get_variables]{bs_get_variables()}},
\code{\link[=builtin_themes]{builtin_themes()}},
\code{\link[=theme_version]{theme_version()}},
\code{\link[=versions]{versions()}}
}
\concept{Bootstrap theme utility functions}
