% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_variables.R
\name{theme_variables}
\alias{theme_variables}
\alias{theme_variables_clear}
\title{Sets SASS variables globally}
\usage{
theme_variables(...)

theme_variables_clear()
}
\arguments{
\item{...}{Variables to be set, using \code{name = value}; or, a single unnamed
argument which is a named list of variables to set. To specify a name that
uses characters that aren't valid for R identifiers, wrap that name in
quotes, double quotes, or backticks (see examples section below).}
}
\value{
If no arguments are passed, a list of all set options is returned. If
arguments are passed, a list containing the previous values of the newly
set options is returned invisibly.
}
\description{
Set SASS variables that will be respected by \code{\link[=bs4_sass]{bs4_sass()}} (by default).
}
\examples{
library(htmltools)

theme_variables(primary = "orange", "body-bg" = "#EEEEEE")
theme_variables("font-family-base" = "monospace", "font-size-base" = "1.4rem")
theme_variables("btn-padding-y" = ".16rem")
theme_variables("btn-padding-x" = "2rem")
theme_variables("border-radius" = 0, "border-radius-lg" = 0, "border-radius-sm" = 0)

browsable(tags$body(
  bs4_sass(),
  tags$a(class = "btn btn-primary", href = "#", role = "button", "Hello")
))

}
