% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootswatch.R
\name{theme_version}
\alias{theme_version}
\title{Obtain a theme's Bootstrap version}
\usage{
theme_version(theme = bs_global_get())
}
\arguments{
\item{theme}{a \code{\link[=bs_theme]{bs_theme()}} object.}
}
\description{
Obtain a theme's Bootstrap version
}
