% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{theme_variables}
\alias{theme_variables}
\alias{theme_variables_clear}
\alias{theme_layer}
\alias{theme_layer_bs3compat}
\alias{theme_layer_bootswatch}
\title{Bootstrap theming helpers}
\usage{
theme_variables(...)

theme_variables_clear()

theme_layer(pre = "", post = "", deps = NULL)

theme_layer_bs3compat()

theme_layer_bootswatch(theme = "", version = version_latest())
}
\arguments{
\item{...}{Variables to be set, using \code{name = value}; or, a single unnamed
argument which is a named list of variables to set. To specify a name that
uses characters that aren't valid for R identifiers, wrap that name in
quotes, double quotes, or backticks (see examples section below).}

\item{pre}{A suitable \code{\link[sass:as_sass]{sass::as_sass()}} \code{input}.}

\item{post}{A suitable \code{\link[sass:as_sass]{sass::as_sass()}} \code{input}.}

\item{deps}{An HTML dependency (or a list of them).}

\item{theme}{a Bootswatch theme name.}

\item{version}{the major version.}
}
\value{
If no arguments are passed, a list of all set options is returned. If
arguments are passed, a list containing the previous values of the newly
set options is returned invisibly.
}
\description{
Use \code{theme_variables()} to set Bootstrap SASS variables globally so
that \code{\link[=bs_sass]{bs_sass()}}/\code{\link[=bs_sass_partial]{bs_sass_partial()}} can respect them at runtime.
Use \code{theme_layer()} to provide \code{\link[sass:sass]{sass::sass()}} to place both before and after
Bootstrap's SASS in \code{\link[=bs_sass]{bs_sass()}}/\code{\link[=bs_sass_partial]{bs_sass_partial()}}.
}
\examples{
library(htmltools)

theme_variables(primary = "orange", "body-bg" = "#EEEEEE")
theme_variables("font-family-base" = "monospace", "font-size-base" = "1.4rem")
theme_variables("btn-padding-y" = ".16rem")
theme_variables("btn-padding-x" = "2rem")
theme_variables("border-radius" = 0, "border-radius-lg" = 0, "border-radius-sm" = 0)

browsable(tags$body(
  bs_sass(),
  tags$a(class = "btn btn-primary", href = "#", role = "button", "Hello")
))

}
\references{
\url{https://getbootstrap.com/docs/4.4/getting-started/theming/}
}
