% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbar.R
\name{toolbar_input_button}
\alias{toolbar_input_button}
\title{Add toolbar button input}
\usage{
toolbar_input_button(
  id,
  label = NULL,
  icon = NULL,
  disabled = FALSE,
  border = FALSE,
  ...
)
}
\arguments{
\item{id}{The \code{input} slot that will be used to access the value.}

\item{label}{The label to display in the button.
(One of icon or label must be supplied.)}

\item{icon}{An icon to display in the button.
(One of icon or label must be supplied.)}

\item{disabled}{If \code{TRUE}, the button will not be clickable.
Use \code{updateActionButton()} to dynamically enable/disable the button.}

\item{border}{Whether to show a border around the button.}

\item{...}{UI elements for the button.}
}
\value{
Returns a button suitable for use in a toolbar.
}
\description{
A button designed to fit well in small places such as toolbars.
}
