% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip}
\alias{tooltip}
\alias{tooltip_toggle}
\alias{tooltip_update}
\title{Add a tooltip to a UI element}
\usage{
tooltip(
  ...,
  body = "Tooltip",
  placement = c("auto", "top", "right", "bottom", "left"),
  html = FALSE,
  sanitize = FALSE,
  id = NULL
)

tooltip_toggle(id, show = NULL, session = get_current_session())

tooltip_update(id, body = NULL, session = get_current_session())
}
\arguments{
\item{...}{Unnamed arguments can be any valid child of an \link[htmltools:builder]{htmltools tag}. Named arguments become HTML attributes on returned
UI element. Attributes starting with \verb{data-bs-*} may be supplied to further
customize \href{https://getbootstrap.com/docs/5.3/components/tooltips/}{tooltip options}.}

\item{body}{Content to display in the tooltip.}

\item{placement}{The placement of the tooltip relative to the target element.}

\item{html}{Whether to treat \code{body} as HTML. WARNING: setting this \code{TRUE}
when the \code{body} is a function of user \code{inputs} is dangerous and not
recommended.}

\item{sanitize}{Whether to sanitize HTML (only relevant when \code{html = TRUE}).
This can be useful if \code{body} is a function of user \code{inputs}, but should
also be treated as HTML.}

\item{id}{a character string that matches an existing tooltip id.}

\item{show}{Whether to show (\code{TRUE}) or hide (\code{FALSE}) the tooltip. The
default (\code{NULL}) will show if currently hidden and hide if currently shown.}

\item{session}{A Shiny session object (the default should almost always be
used).}
}
\description{
Tooltips are useful for showing additional information when focusing (or
hovering over) a UI element.
}
\details{
If multiple UI elements are provided to \code{...}, the last element is
used as the tooltip's target. If the target should contain multiple elements,
then wrap those elements in a \code{\link[=span]{span()}} or \code{\link[=div]{div()}}.
}
\section{Functions}{
\itemize{
\item \code{tooltip()}: Add a tooltip to a UI element

\item \code{tooltip_toggle()}: Programmatically show/hide a tooltip.

\item \code{tooltip_update()}: Update the \code{body} of a tooltip.

}}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

tooltip(
  shiny::actionButton("btn", "A button"),
  body = "A message"
)

card(
  card_header(
    tooltip(
      span("Card title ", bsicons::bs_icon("question-circle-fill")),
      body = "Additional info",
      placement = "right"
    )
  ),
  "Card body content..."
)
\dontshow{\}) # examplesIf}
}
\references{
\url{https://getbootstrap.com/docs/5.3/components/tooltips/}
}
