% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input_date.R
\name{update_date}
\alias{update_date}
\title{Change the value of a date input on the client}
\usage{
update_date(
  id,
  ...,
  label = NULL,
  value = NULL,
  min = NULL,
  max = NULL,
  session = get_current_session()
)
}
\arguments{
\item{...}{Ignored, included for future expansion.}

\item{label}{The label to set for the input object.}

\item{value}{The starting date. Either a Date object, or a string in
\code{yyyy-mm-dd} format. If NULL (the default), will use the current date
in the client's time zone.}

\item{min}{The minimum allowed date. Either a Date object, or a string in
\code{yyyy-mm-dd} format.}

\item{max}{The maximum allowed date. Either a Date object, or a string in
\code{yyyy-mm-dd} format.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}
}
\description{
Change the value of a date input on the client
}
\details{
The input updater functions send a message to the client, telling it to
change the settings of an input object. The messages are collected and sent
after all the observers (including outputs) have finished running.

The syntax of these functions is similar to the functions that created the
inputs in the first place. For example, \code{\link[shiny]{numericInput}()} and
\code{updateNumericInput()} take a similar set of arguments.

Any arguments with NULL values will be ignored; they will not result in any
changes to the input object on the client.

For \code{\link[shiny]{radioButtons}()}, \code{\link[shiny]{checkboxGroupInput}()} and
\code{\link[shiny]{selectInput}()}, the set of choices can be cleared by using
\code{choices=character(0)}. Similarly, for these inputs, the selected item
can be cleared by using \code{selected=character(0)}.
}
\seealso{
Other Shiny update aliases: 
\code{\link{update_action_button}()},
\code{\link{update_action_link}()},
\code{\link{update_checkbox}()},
\code{\link{update_date_range}()},
\code{\link{update_select}()},
\code{\link{update_selectize}()},
\code{\link{update_slider}()}
}
\concept{Shiny update aliases}
