% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input_numeric.R
\name{update_numeric}
\alias{update_numeric}
\title{Change the value of a number input on the client}
\usage{
update_numeric(
  id,
  ...,
  label = NULL,
  value = NULL,
  min = NULL,
  max = NULL,
  step = NULL,
  session = get_current_session()
)
}
\arguments{
\item{id}{An input id.}

\item{...}{Ignored, included for future expansion.}

\item{label}{The label to set for the input object.}

\item{value}{Initial value.}

\item{min}{Minimum allowed value}

\item{max}{Maximum allowed value}

\item{step}{Interval to use when stepping between min and max}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}
}
\description{
Change the value of a number input on the client
}
\details{
The input updater functions send a message to the client, telling it to
change the settings of an input object. The messages are collected and sent
after all the observers (including outputs) have finished running.

The syntax of these functions is similar to the functions that created the
inputs in the first place. For example, \code{\link[shiny]{numericInput}()} and
\code{updateNumericInput()} take a similar set of arguments.

Any arguments with NULL values will be ignored; they will not result in any
changes to the input object on the client.

For \code{\link[shiny]{radioButtons}()}, \code{\link[shiny]{checkboxGroupInput}()} and
\code{\link[shiny]{selectInput}()}, the set of choices can be cleared by using
\code{choices=character(0)}. Similarly, for these inputs, the selected item
can be cleared by using \code{selected=character(0)}.
}
\seealso{
Other Shiny update aliases: 
\code{\link{update_action_button}()},
\code{\link{update_action_link}()},
\code{\link{update_checkbox}()},
\code{\link{update_checkbox_group}()},
\code{\link{update_date}()},
\code{\link{update_date_range}()},
\code{\link{update_radio_buttons}()},
\code{\link{update_select}()},
\code{\link{update_selectize}()},
\code{\link{update_slider}()},
\code{\link{update_text}()},
\code{\link{update_text_area}()}
}
\concept{Shiny update aliases}
