% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input_radio_buttons.R
\name{update_radio_buttons}
\alias{update_radio_buttons}
\title{Change the value of a radio input on the client}
\usage{
update_radio_buttons(
  id,
  ...,
  label = NULL,
  choices = NULL,
  selected = NULL,
  inline = FALSE,
  choiceNames = NULL,
  choiceValues = NULL,
  session = get_current_session()
)
}
\arguments{
\item{id}{An input id.}

\item{...}{Ignored, included for future expansion.}

\item{label}{The label to set for the input object.}

\item{choices}{List of values to select from (if elements of the list are
named then that name rather than the value is displayed to the user). If
this argument is provided, then \code{choiceNames} and \code{choiceValues} must not
be provided, and vice-versa. The values should be strings; other types
(such as logicals and numbers) will be coerced to strings.}

\item{selected}{The initially selected value. If not specified, then it
defaults to the first item in \code{choices}. To start with no items selected,
use \code{character(0)}.}

\item{inline}{If \code{TRUE}, render the choices inline (i.e. horizontally)}

\item{choiceNames, choiceValues}{List of names and values, respectively, that
are displayed to the user in the app and correspond to the each choice (for
this reason, \code{choiceNames} and \code{choiceValues} must have the same length).
If either of these arguments is provided, then the other \emph{must} be provided
and \code{choices} \emph{must not} be provided. The advantage of using both of these
over a named list for \code{choices} is that \code{choiceNames} allows any type of UI
object to be passed through (tag objects, icons, HTML code, ...), instead
of just simple text. See Examples.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}
}
\description{
Change the value of a radio input on the client
}
\details{
The input updater functions send a message to the client, telling it to
change the settings of an input object. The messages are collected and sent
after all the observers (including outputs) have finished running.

The syntax of these functions is similar to the functions that created the
inputs in the first place. For example, \code{\link[shiny]{numericInput}()} and
\code{updateNumericInput()} take a similar set of arguments.

Any arguments with NULL values will be ignored; they will not result in any
changes to the input object on the client.

For \code{\link[shiny]{radioButtons}()}, \code{\link[shiny]{checkboxGroupInput}()} and
\code{\link[shiny]{selectInput}()}, the set of choices can be cleared by using
\code{choices=character(0)}. Similarly, for these inputs, the selected item
can be cleared by using \code{selected=character(0)}.
}
\section{Aliased from Shiny}{
 This function is an alias for \code{shiny::updateRadioButtons()} and is included to maintain more consistent naming conventions in Shiny apps that use \pkg{bslib}. The documentation may still refer to the original function names. You can replace \code{shiny::updateRadioButtons()} with \code{update_radio_buttons()}.
}

\seealso{
\code{\link[=input_radio_buttons]{input_radio_buttons()}} to create a radio button input.

Other Shiny update aliases: 
\code{\link{update_action_button}()},
\code{\link{update_action_link}()},
\code{\link{update_checkbox}()},
\code{\link{update_checkbox_group}()},
\code{\link{update_date}()},
\code{\link{update_date_range}()},
\code{\link{update_numeric}()},
\code{\link{update_select}()},
\code{\link{update_selectize}()},
\code{\link{update_slider}()},
\code{\link{update_text}()},
\code{\link{update_text_area}()}
}
\concept{Shiny update aliases}
