% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value-box.R
\name{value_box}
\alias{value_box}
\title{A card for displaying a summary of information}
\usage{
value_box(
  title,
  value,
  ...,
  showcase = NULL,
  showcase_layout = c("top-right", "left-center"),
  full_screen = FALSE,
  class = NULL
)
}
\arguments{
\item{title}{a \code{\link[htmltools:builder]{htmltools::tag()}} child to display above \code{value}.}

\item{value}{a \code{\link[htmltools:builder]{htmltools::tag()}} child to display below \code{title}.}

\item{...}{Unnamed arguments may be any \code{\link[htmltools:builder]{htmltools::tag()}} children
to display below \code{value} (if these children are not already
\code{\link[=is.card_item]{is.card_item()}}, then they are wrapped in a \code{\link[=card_body]{card_body()}}).
Named arguments become attributes on the containing element.}

\item{showcase}{a \code{\link[htmltools:builder]{htmltools::tag()}} child to showcase (e.g., a
\code{\link[bsicons:bs_icon]{bsicons::bs_icon()}}, a \code{\link[plotly:plotlyOutput]{plotly::plotlyOutput()}}, etc).}

\item{showcase_layout}{where to display the showcase relative to the other
content.}

\item{full_screen}{If \code{TRUE}, an icon will appear when hovering over the card
body. Clicking the icon expands the card to fit viewport size. Consider
pairing this feature with \code{\link[=card_body_fill]{card_body_fill()}} to get output that responds to
changes in the size of the card.}

\item{class}{utility classes for customizing the appearance of the summary
card. Use \verb{bg-*} and \verb{text-*} classes (e.g, \code{"bg-danger"} and
\code{"text-light"}) to customize the background/foreground colors.}
}
\description{
A card for displaying a summary of information
}
\examples{

library(htmltools)

if (interactive()) {
  value_box(
    "KPI Title",
    HTML("$1 <i>Billion</i> Dollars"),
    span(
      bsicons::bs_icon("arrow-up"),
      " 30\% VS PREVIOUS 30 DAYS"
     ),
    showcase = bsicons::bs_icon("piggy-bank"),
    class = "bg-success"
  )
}
}
