% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version-default.R
\name{versions}
\alias{versions}
\alias{version_default}
\title{Available Bootstrap versions}
\usage{
versions()

version_default()
}
\value{
Returns a list of the Bootstrap versions available.
}
\description{
Available Bootstrap versions
}
\seealso{
Other Bootstrap theme utility functions:
\code{\link[=bootswatch_themes]{bootswatch_themes()}},
\code{\link[=bs_get_variables]{bs_get_variables()}},
\code{\link[=builtin_themes]{builtin_themes()}},
\code{\link[=theme_bootswatch]{theme_bootswatch()}},
\code{\link[=theme_version]{theme_version()}}
}
\concept{Bootstrap theme utility functions}
