% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{web_font}
\alias{web_font}
\alias{gfont}
\title{Import web fonts into a \code{\link[=bs_theme]{bs_theme()}}}
\usage{
web_font(name, url, ...)

gfont(name, weight = NULL, local = TRUE, cache = TRUE)
}
\arguments{
\item{name}{font family name}

\item{url}{a URL pointing the font file}

\item{...}{other CSS properties to place on the \verb{@font-face} definition.}

\item{weight}{A Google Font axis weight definition.}

\item{local}{whether or not download and bundle local (woff) font files.}

\item{cache}{whether or not to cache local font files (only takes effect when
\code{local = TRUE}).}
}
\description{
When used with any of the main font settings in \code{\link[=bs_theme]{bs_theme()}} (e.g.
\code{base_font}, \code{code_font}, \code{heading_font}), these font helpers ensure
relevant font files sources are included with the theme. Google Fonts
may be used either offline or online, but Web Fonts require an
internet connection.
}
\examples{
theme <- bs_theme(base_font = gfont("Pacifico"))
if (interactive()) {
  bs_theme_preview(theme)
}
m <- web_font("Yellowtail", url = "https://fonts.googleapis.com/css?family=Yellowtail")
theme <- bs_theme(base_font = m)
if (interactive()) {
  bs_theme_preview(theme)
}

}
\references{
\url{https://developer.mozilla.org/en-US/docs/Learn/CSS/Styling_text/Web_fonts}

\url{https://developers.google.com/fonts/docs/css2}
}
