% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion.R
\name{accordion_select}
\alias{accordion_select}
\alias{accordion_insert}
\alias{accordion_remove}
\alias{accordion_replace}
\title{Dynamically update accordions}
\usage{
accordion_select(id, selected, close = TRUE, session = get_current_session())

accordion_insert(
  id,
  panel,
  target = NULL,
  position = c("after", "before"),
  session = get_current_session()
)

accordion_remove(id, target, session = get_current_session())

accordion_replace(
  id,
  target,
  ...,
  title = NULL,
  value = NULL,
  icon = NULL,
  session = get_current_session()
)
}
\arguments{
\item{id}{an character string that matches an existing \code{\link[=accordion]{accordion()}}'s \code{id}.}

\item{selected}{a character string used to identify a particular \code{\link[=accordion_panel]{accordion_panel()}}.}

\item{close}{whether to close non-\code{selected} \code{\link[=accordion_panel]{accordion_panel()}}s. Note that \code{close=FALSE} won't be respected when the \code{\link[=accordion]{accordion()}} has \code{autoclose=TRUE}.}

\item{session}{a shiny session object (the default should almost always be used).}

\item{panel}{an \code{\link[=accordion_panel]{accordion_panel()}}.}

\item{target}{The \code{value} of an existing \code{panel} to insert next to. If removing: the \code{value} of the \code{\link[=accordion_panel]{accordion_panel()}} to remove.}

\item{position}{Should \code{panel} be added before or after the target?}

\item{...}{Named arguments become attributes on the \verb{<div class="accordion">}
element. Unnamed arguments should be \code{accordion_panel()}s.}

\item{title}{A title to appear in the \code{accordion_panel()}'s header.}

\item{value}{A character string that uniquely identifies this panel.}

\item{icon}{A \link[htmltools:builder]{htmltools::tag} child (e.g., \code{\link[bsicons:bs_icon]{bsicons::bs_icon()}}) which is positioned just before the \code{title}.}
}
\description{
Functions for dynamically updating nav containers (e.g., select, insert, and
remove nav items). These functions require an \code{id} on the nav container to be
specified and must be called within an active Shiny session.
}
