% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{as_fill_carrier}
\alias{as_fill_carrier}
\alias{as_fillable}
\alias{as_fill}
\alias{undo_fill}
\alias{is_fill_carrier}
\alias{is_fillable}
\alias{is_fill}
\title{Coerce a UI element to be a fill carrier / item / fillable container}
\usage{
as_fill_carrier(x, ..., class = NULL, style = NULL, .selector = NULL)

as_fillable(
  x,
  ...,
  min_height = NULL,
  max_height = NULL,
  gap = NULL,
  class = NULL,
  style = NULL,
  .selector = NULL
)

as_fill(
  x,
  ...,
  min_height = NULL,
  max_height = NULL,
  class = NULL,
  style = NULL,
  .selector = NULL
)

undo_fill(x)

is_fill_carrier(x)

is_fillable(x)

is_fill(x)
}
\arguments{
\item{x}{a \code{\link[htmltools:builder]{htmltools::tag()}}.}
}
\description{
TODO: link to filling layout article.
}
