% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_sass.R
\name{bootswatch_themes}
\alias{bootswatch_themes}
\title{List all Bootswatch themes}
\usage{
bootswatch_themes(version = version_latest(), full_path = FALSE)
}
\arguments{
\item{version}{the major version of Bootswatch.}

\item{full_path}{whether to return a path to the installed theme.}
}
\description{
List all Bootswatch themes
}
