% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-theme-preset-bootswatch.R
\name{bootswatch_themes}
\alias{bootswatch_themes}
\title{Obtain a list of all available bootswatch themes.}
\usage{
bootswatch_themes(version = version_default(), full_path = FALSE)
}
\arguments{
\item{version}{The major version of Bootswatch.}

\item{full_path}{Whether to return a path to the installed theme.}
}
\value{
Returns a character vector of Bootswatch themes.
}
\description{
Obtain a list of all available bootswatch themes.
}
\seealso{
Other Bootstrap theme utility functions: 
\code{\link{bs_get_variables}()},
\code{\link{builtin_themes}()},
\code{\link{theme_bootswatch}()},
\code{\link{theme_version}()},
\code{\link{versions}()}
}
\concept{Bootstrap theme utility functions}
