% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootswatch.R
\name{bootswatch_themes}
\alias{bootswatch_themes}
\title{Obtain a list of all available bootswatch themes.}
\usage{
bootswatch_themes(version = version_default(), full_path = FALSE)
}
\arguments{
\item{version}{the major version of Bootswatch.}

\item{full_path}{whether to return a path to the installed theme.}
}
\value{
a character vector of Bootswatch themes.
}
\description{
Obtain a list of all available bootswatch themes.
}
