% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{breakpoints}
\alias{breakpoints}
\title{Define breakpoint values}
\usage{
breakpoints(..., sm = NULL, md = NULL, lg = NULL)
}
\arguments{
\item{...}{Other breakpoints (e.g., \code{xl}).}

\item{sm}{Values to apply at the \code{sm} breakpoint.}

\item{md}{Values to apply at the \code{md} breakpoint.}

\item{lg}{Values to apply at the \code{lg} breakpoint.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A generic constructor for responsive breakpoints.
}
\examples{
breakpoints(sm = c(4, 4, 4), md = c(3, 3, 6), lg = c(-2, 8, -2))

}
\references{
\url{https://getbootstrap.com/docs/5.3/layout/breakpoints/}
}
\seealso{
\code{\link[=layout_columns]{layout_columns()}}
}
