% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs4_sass.R
\name{bs4_sass}
\alias{bs4_sass}
\alias{bs4_scss_required}
\alias{bs4_scss_file}
\title{Compile Bootstrap 4 SASS with (optional) theming}
\usage{
bs4_sass(variables = theme_variables(), theme = bs4_theme_bs3compat(),
  bootstrap_scss = bs4_scss_file("bootstrap.scss"),
  options = sass_options())

bs4_scss_required()

bs4_scss_file(path)
}
\arguments{
\item{variables}{A list of SASS variables to include prior to any bootstrap or
\code{theme} sass. To ignore any existing variables (set by \code{\link[=theme_variables]{theme_variables()}}) at
runtime, set to \code{NULL}.}

\item{theme}{a \code{\link[=bs4_theme]{bs4_theme()}} object. Use this argument to define your own custom
theme and/or use a pre-defined theme (e.g., \code{\link[=bs4_theme_bootswatch]{bs4_theme_bootswatch()}}).}

\item{bootstrap_scss}{bootstrap scss files to compile \code{variables} and \code{theme} against.
By default, all of bootstrap is included. Use \code{bs4_scss_required()} to compile against
'core' parts of bootstrap. Use \code{bs4_scss_file()} to fully customize what files to import.}

\item{options}{Compiler options for SASS, see \code{\link[sass:sass_options]{sass::sass_options()}}.}
}
\description{
Obtain Bootstrap (and it's dependencies) as an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
By default, \code{bs4_sass()} returns vanilla Bootstrap, but it also provides
various ways to theme (see examples below).
}
\examples{
library(htmltools)

# Vanilla Bootstrap 4
button <- tags$a(class = "btn btn-primary", href = "#", role = "button", "Hello")
browsable(tags$body(bs4_sass(), button))

# Use a bootswatch theme
bs4_minty <- bs4_sass(theme = bs4_theme_bootswatch("minty"))
browsable(tags$body(bs4_minty, button))

# Set bootstrap SASS variables (globally)
theme_variables(primary = "orange", "body-bg" = "#EEEEEE")
theme_variables("font-family-base" = "monospace", "font-size-base" = "1.4rem")
theme_variables("btn-padding-y" = ".16rem")
theme_variables("btn-padding-x" = "2rem")
theme_variables("border-radius" = 0, "border-radius-lg" = 0, "border-radius-sm" = 0)
browsable(tags$body(bs4_sass(), button))

# Use a custom theme
person <- function(name, title, company) {
  tags$div(
    class = "person",
    h3(class = "name", name),
    div(class = "title", title),
    div(class = "company", company)
  )
}
person_scss <- sass::sass_file(system.file("custom", "person.scss", package = "bootscss"))
browsable(tags$body(
  bs4_sass(theme = bs4_theme(post = person_scss)),
  person("Andrew Carnegie", "Owner", "Carnegie Steel Company"),
  person("John D. Rockefeller", "Chairman", "Standard Oil")
))

}
\references{
\url{https://getbootstrap.com/docs/4.3/getting-started/theming/}
}
\seealso{
\code{\link[=bs4_theme]{bs4_theme()}}, \code{\link[=theme_variables]{theme_variables()}}
}
