% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{bs_page}
\alias{bs_page}
\title{Creat a Bootstrap page}
\usage{
bs_page(..., title = NULL, theme = bs_theme(), lang = NULL)
}
\description{
Alias for \code{\link[shiny:bootstrapPage]{shiny::bootstrapPage()}} with \code{theme} defaulting to a newer version
of Bootstrap.
}
