% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_sass.R
\name{bs_sass}
\alias{bs_sass}
\alias{bs_sass_partial}
\title{Compile Bootstrap 4 (or 3) SASS with (optional) theming}
\usage{
bs_sass(
  ...,
  variables = theme_variables(),
  bootswatch = NULL,
  version = c("4-3", "4", "3"),
  jquery = jquerylib::jquery_core(3),
  options = sass_options()
)

bs_sass_partial(
  ...,
  variables = theme_variables(),
  bootswatch = NULL,
  version = 4,
  options = sass_options()
)
}
\arguments{
\item{...}{a collection of \code{\link[sass:sass]{sass::sass()}} to include after Bootstrap's SASS.
Use \code{\link[=bs_theme]{bs_theme()}} if you need to place \code{\link[sass:sass]{sass::sass()}} both before and after Bootstrap's SASS.}

\item{variables}{A list of SASS variables to include prior to any other SASS.
To ignore any existing variables (set by \code{\link[=theme_variables]{theme_variables()}}) at runtime,
set this to \code{NULL}.}

\item{bootswatch}{The name of a bootswatch theme.
See \code{\link[=bootswatch_themes]{bootswatch_themes()}} for a list of possible names.}

\item{version}{The major version of Bootstrap to use. A value of
\code{'4-3'} means Bootstrap 4, but with additional CSS/JS to support
BS3 style markup in BS4.}

\item{jquery}{See \code{\link[jquerylib:jquery_core]{jquerylib::jquery_core()}}.}

\item{options}{Compiler options for Sass. Please specify options using
\code{\link[sass]{sass_options}}.}
}
\description{
Use \code{bs_sass()} to obtain all of Bootstrap and it's dependencies as a list
of \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} objects.
Use \code{bs_sass_partial()} if you can assume Bootstrap already exists in your
project, but you want to leverage Bootstrap utilities to generate additional
CSS (as a string that can be included as a \verb{<style>} tag via \code{tags$style(css)}).
}
\examples{
library(htmltools)

button <- tags$a(class = "btn btn-primary", href = "#", role = "button", "Hello")

preview_button <- function(x) {
  browsable(tags$body(x, button))
}

# Vanilla Bootstrap 4
preview_button(bs_sass())

# Vanilla Bootstrap 3
preview_button(bs_sass(version = 3))

# Bootswatch 4 theme minty
preview_button(bs_sass(bootswatch = "minty"))

# Bootswatch 4 theme sketchy
preview_button(bs_sass(bootswatch = "sketchy"))

# Set bootstrap SASS variables (globally)
theme_variables(primary = "orange", "body-bg" = "#EEEEEE")
theme_variables("font-family-base" = "monospace", "font-size-base" = "1.4rem")
theme_variables("btn-padding-y" = ".16rem")
theme_variables("btn-padding-x" = "2rem")
theme_variables("border-radius" = 0, "border-radius-lg" = 0, "border-radius-sm" = 0)
preview_button(bs_sass())

# Use a custom theme
person <- function(name, title, company) {
  tags$div(
    class = "person",
    h3(class = "name", name),
    div(class = "title", title),
    div(class = "company", company)
  )
}
person_scss <- sass::sass_file(system.file("custom", "person.scss", package = "bootstraplib"))
browsable(tags$body(
  bs_sass(person_scss),
  person("Andrew Carnegie", "Owner", "Carnegie Steel Company"),
  person("John D. Rockefeller", "Chairman", "Standard Oil")
))

}
\references{
\url{https://getbootstrap.com/docs/4.3/getting-started/theming/}
}
\seealso{
\code{\link[=bs_theme]{bs_theme()}}, \code{\link[=theme_variables]{theme_variables()}}
}
