% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-base-colors.R
\name{bs_theme_base_colors}
\alias{bs_theme_base_colors}
\title{Customize the Bootstrap theme based on two to four key colors}
\usage{
bs_theme_base_colors(bg = "#FFFFFF", fg = "#000000")
}
\arguments{
\item{bg}{A color string for the background, in any format
\code{\link[htmltools:parseCssColors]{htmltools::parseCssColors()}} can understand.}

\item{fg}{A color string for the background.}
}
\description{
\code{bs_theme_base_colors} uses a background color and a foreground color to
rewrite the basic palette of the current Bootstrap theme, affecting almost
every built-in Bootstrap component.
}
\section{Implementation notes}{


\code{bs_theme_base_colors} supports both Bootstrap 3 and 4 (and 4+3), but the
implementation differs greatly.

For Bootstrap 4, the vast majority of default colors are directly or
indirectly based on the \verb{$black}, \verb{$white}, and \verb{$gray-100} through
\verb{$gray-900} variables; or on the theme colors (primary, secondary, danger,
warning, info, etc.). \code{bs_theme_base_colors} sets \verb{$white} to the \code{bg} color,
\verb{$black} to the \code{fg} color, and interpolates the grays between them.

For Bootstrap 3, a similar set of \verb{$black}, \verb{$white}, and \verb{$gray-darker}
through \verb{$gray-lighter} variables exist, and these are populated using the
same strategy as with Bootstrap 4. However, unlike Bootstrap 4, in Bootstrap
3 many of the default colors are hard-coded hex colors, that also happen to
be shades of gray. \code{bs_theme_base_colors} overrides these hard-coded values
with colors interpolated between \code{bg} and \code{fg}.
}

\examples{

bs_theme_new("4+3")
bs_theme_base_colors(bg = "#000060", fg = "skyblue")

# You can apply further customizations here if desired, e.g.:
bs_theme_accent_colors(primary = "orange", secondary = "silver")

if (interactive()) {
  bs_theme_preview(with_themer = FALSE)
}

}
\seealso{
Other customizations: 
\code{\link{bs_theme_accent_colors}()},
\code{\link{bs_theme_fonts}()}
}
\concept{customizations}
