% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-preview.R
\name{bs_theme_preview}
\alias{bs_theme_preview}
\title{Preview the currently set theme}
\usage{
bs_theme_preview(
  ...,
  with_themer = TRUE,
  auto_theme = rlang::is_installed("thematic")
)
}
\arguments{
\item{...}{passed along to \code{\link[shiny:runApp]{shiny::runApp()}}}

\item{with_themer}{whether or not to run the app with \code{\link[=run_with_themer]{run_with_themer()}}.}

\item{auto_theme}{whether or not to ensure plot auto theming is enabled (requires thematic package).}
}
\description{
Launches an example shiny app via \code{run_with_themer()} and \code{bootstrap()}.
Useful for getting a quick preview of the current theme setting as
well as an interactive GUI for tweaking some of the main theme settings.
}
\details{
The app that this launches is subject to change.
}
\examples{

bs_theme_new()
bs_theme_base_colors("#6c757d", "white")
bs_theme_accent_colors("orange")
if (interactive()) bs_theme_preview()

}
\seealso{
\code{\link[=run_with_themer]{run_with_themer()}}
}
