% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{as.card_item}
\alias{as.card_item}
\alias{is.card_item}
\alias{card_body}
\alias{card_header}
\alias{card_heading}
\alias{card_title}
\alias{card_footer}
\alias{card_spacer}
\alias{card_list}
\alias{card_plot_output}
\title{Card components}
\usage{
as.card_item(x)

is.card_item(x)

card_body(..., class = NULL, padding = c("x", "y"), stretch = FALSE)

card_header(..., class = NULL)

card_heading(..., class = NULL, container = htmltools::tags$h6)

card_title(..., class = NULL)

card_footer(..., class = NULL)

card_spacer(...)

card_list(...)

card_plot_output(
  outputId,
  click = NULL,
  dblclick = NULL,
  hover = NULL,
  brush = NULL,
  height = NULL,
  stretch = TRUE
)
}
\description{
This topic describes various components that are intended to go
directly inside of a \code{\link[=card]{card()}}. These components can be used in combination;
for example, a single card could contain a \code{card_header()}, multiple
\code{card_body()}s, a \code{card_list()}, and finally a \code{card_footer()}.
}
