% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card_grid}
\alias{card_grid}
\title{Card grid layout}
\usage{
card_grid(
  ...,
  class = NULL,
  card_width = 1/4,
  gap = NULL,
  heights_equal = c("all", "row"),
  fixed_width = FALSE
)
}
\arguments{
\item{...}{Named arguments become attributes on the containing
\link[htmltools:builder]{htmltools::tag} element. Unnamed arguments are children, and should be
instances of \code{\link[=card]{card()}} (and/or \code{\link[=navs_tab_card]{navs_tab_card()}}/\code{\link[=navs_pill_card]{navs_pill_card()}}).}

\item{class}{Additional CSS classes to include on the grid div.}

\item{card_width}{The desired width of each card, which can be any of the
following: * A (unit-less) number between 0 and 1. * This should be
specified as \code{1/num}, where \code{num} represents the number of desired columns.
\itemize{
\item A \link[htmltools:validateCssUnit]{CSS length unit} * Either the minimum
(when \code{fixed_width=FALSE}) or fixed width (\code{fixed_width=TRUE}). * \code{NULL} *
Allows power users to set the \code{grid-template-columns} CSS property
manually, either via a \code{style} attribute or a CSS stylesheet.
}}

\item{gap}{A \link[htmltools:validateCssUnit]{CSS length unit} that sets the
amount of spacing between cards. If \code{NULL}, the value of Bootstrap's
\verb{$spacer} Sass variable is used.}

\item{heights_equal}{If \code{"all"} (the default), every card in every row of the
grid will have the same height. If \code{"row"}, then every card in \emph{each} row
of the grid will have the same height, but heights may vary between rows.}

\item{fixed_width}{Whether or not to interpret the \code{card_width} as a minimum
(\code{fixed_width=FALSE}) or fixed (\code{fixed_width=TRUE}) width when it is a CSS
length unit.}
}
\description{
Layout a collection of cards into a responsive grid. By default,
cards will have the same height/width, but optionally, the heights of each
row (and column) may be different.

To learn more about \code{\link[=card]{card()}}s and card layout options, see \href{https://rstudio.github.io/bslib/articles/cards.html}{this article}.
}
\seealso{
\code{\link[=card]{card()}} for creating a card container.

\code{\link[=as.card_item]{as.card_item()}} for putting stuff inside the card.

\code{\link[=nabs_tab_card]{nabs_tab_card()}} \code{\link[=nabs_pill_card]{nabs_pill_card()}} for placing navigation in cards.
}
