% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card_grid}
\alias{card_grid}
\title{Card grid}
\usage{
card_grid(
  ...,
  class = NULL,
  card_width = 1/4,
  gap = NULL,
  heights_equal = c("all", "row"),
  fixed_width = FALSE
)
}
\arguments{
\item{...}{Named arguments become attributes on the \verb{<div class="card">}
element. Unnamed arguments are children, and should be instances of
\code{\link[=card]{card()}}.}

\item{class}{Additional CSS classes to include on the grid div.}

\item{card_width}{The desired width of each card. This can be any \link[htmltools:validateCssUnit]{CSS length unit}, in which case \code{card_grid} will fit as
many cards into each row without being smaller than that width.
Alternatively, you can specify a number of cards per row by passing
\code{1/num}, for example \code{card_width = 1/3} for three cards per row. You can
also set this to \code{NULL} and set the \code{grid-template-columns} CSS property
manually, either via a \code{style} attribute or a CSS stylesheet.}

\item{gap}{A \link[htmltools:validateCssUnit]{CSS length unit} that sets the
amount of spacing between cards. If \code{NULL}, the value of Bootstrap's
\verb{$spacer} Sass variable is used.}

\item{heights_equal}{If \code{"all"} (the default), every card in every row of the
grid will have the same height. If \code{"row"}, then every card in \emph{each} row
of the grid will have the same height, but heights may vary between rows.}

\item{fixed_width}{When \code{card_width} is a CSS length (as opposed to a
fraction), the default behavior is to let cards be that wide \emph{or larger},
so that they fill all of the horizontal space that's available. Set this
argument to \code{TRUE} to ensure that cards are exactly the specified width,
even if this means using less than the total width of the card grid
container.}
}
\description{
Use to create a grid of cards, giving every card in the entire grid the same
height/width (or optionally, every card in \emph{each row} of the grid).
}
