% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{bs_theme_new}
\alias{bs_theme_new}
\alias{bs_theme_clear}
\alias{bs_theme_get}
\alias{bs_theme_set}
\alias{bs_theme_base_colors}
\alias{bs_theme_accent_colors}
\alias{bs_theme_fonts}
\alias{bs_theme_add_variables}
\alias{bs_theme_add}
\alias{bs_theme_get_variables}
\alias{bootstrap}
\alias{bootstrap_sass}
\title{Deprecated functions}
\usage{
bs_theme_new(...)

bs_theme_clear(...)

bs_theme_get(...)

bs_theme_set(...)

bs_theme_base_colors(...)

bs_theme_accent_colors(...)

bs_theme_fonts(...)

bs_theme_add_variables(...)

bs_theme_add(...)

bs_theme_get_variables(...)

bootstrap(...)

bootstrap_sass(...)
}
\description{
These functions have been deprecated but remain for backwards compatibility.
}
\keyword{internal}
