% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{font_face}
\alias{font_face}
\alias{font_link}
\alias{font_google}
\title{Import web fonts into a \code{\link[=bs_theme]{bs_theme()}}}
\usage{
font_face(
  family,
  src,
  weight = NULL,
  display = NULL,
  style = NULL,
  stretch = NULL,
  variant = NULL,
  unicode_range = NULL
)

font_link(family, href)

font_google(family, weight = NULL, local = TRUE, cache = TRUE)
}
\arguments{
\item{family}{font family name.}

\item{src}{}

\item{weight}{A Google Font axis weight definition.}

\item{href}{A hyperlink to a resource containing the font data.}

\item{local}{whether or not download and bundle local (woff) font files.}

\item{cache}{whether or not to cache local font files (only takes effect when
\code{local = TRUE}).}
}
\description{
When used with any of the main font settings in \code{\link[=bs_theme]{bs_theme()}} (e.g.,
\code{base_font}, \code{code_font}, \code{heading_font}), these font objects ensure relevant
font files sources are included with the theme. In order to work,
\code{\link[=web_font]{web_font()}}s requires that the end user has internet access, but \code{gfont()}
may be used offline once the files have been downloaded locally (\code{local = TRUE}).
}
\examples{

# If you have an internet connection, running the following code
# will download, cache, and import the relevant Google Font files
# for local use
theme <- bs_theme(
  base_font = font_google("Fira Sans"),
  code_font = font_google("Fira Code"),
  heading_font = font_google("Fredoka One")
)
if (interactive()) {
  bs_theme_preview(theme)
}

# Three different yet equivalent ways of importing a remotely-hosted Google Font
a <- font_google("Crimson Pro", weight = "200..900", local = FALSE)
b <- font_link("Crimson Pro", href = "https://fonts.googleapis.com/css2?family=Crimson+Pro:wght@200..900")
c <- font_face(
  family = "Crimson Pro",
  style = "normal",
  weight = "200 900",
  src = "url(https://fonts.gstatic.com/s/crimsonpro/v13/q5uDsoa5M_tv7IihmnkabARboYF6CsKj.woff2) format('woff2')"
)
theme <- bs_theme(base_font = c)
if (interactive()) {
  bs_theme_preview(theme)
}

}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/CSS/@font-face}, \url{https://developer.mozilla.org/en-US/docs/Learn/CSS/Styling_text/Web_fonts}

\url{https://developers.google.com/fonts/docs/css2}
}
