% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.bslib_fragment}
\alias{print.bslib_fragment}
\alias{print.bslib_page}
\title{Print a bslib page/fragment}
\usage{
\method{print}{bslib_fragment}(x, ...)

\method{print}{bslib_page}(x, ...)
}
\arguments{
\item{x}{a bslib page/fragment.}

\item{...}{passed along to an underlying print method.}
}
\description{
Print a bslib page/fragment
}
\keyword{internal}
