% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-button-group.R
\name{input_check_buttons}
\alias{input_check_buttons}
\alias{input_radio_buttons}
\alias{update_check_buttons}
\alias{update_radio_buttons}
\title{Create a button group of radio/check boxes}
\usage{
input_check_buttons(id, choices, ..., selected = NULL, gap = 0, theme = NULL)

input_radio_buttons(id, choices, ..., selected = NULL, gap = 0, theme = NULL)

update_check_buttons(
  id,
  choices = NULL,
  selected = NULL,
  session = get_current_session()
)

update_radio_buttons(
  id,
  choices = NULL,
  selected = NULL,
  session = get_current_session()
)
}
\arguments{
\item{theme}{a theme color.}
}
\description{
Use \code{input_check_buttons()} if multiple choices may be selected at once; otherwise, use \code{input_radio_buttons()}
}
