% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check-search.R
\name{input_check_search}
\alias{input_check_search}
\title{A searchable list of checkboxes}
\usage{
input_check_search(
  id,
  choices,
  selected = NULL,
  placeholder = "🔍 Search",
  height = NULL,
  width = NULL
)
}
\arguments{
\item{id}{an input id.}

\item{choices}{a vector/list of choices. If there are names on the on the vector, those names are used as the input value.}

\item{selected}{a vector/list of choices to select by default.}

\item{placeholder}{some text to appear when no search input is provided}

\item{height}{a valid CSS unit for the height of the input.}
}
\description{
A searchable list of checkboxes
}
