% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-input_checkbox.R
\name{input_checkbox}
\alias{input_checkbox}
\title{Checkbox Input Control}
\usage{
input_checkbox(id, label, value = FALSE, ..., width = NULL)
}
\arguments{
\item{id}{An input id.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}

\item{...}{Ignored, included for future expansion.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}
}
\value{
A checkbox control that can be added to a UI definition.
}
\description{
Create a checkbox that can be used to specify logical values.
}
\section{Aliased from Shiny}{
 This function is an alias for \code{shiny::checkboxInput()} and is included to maintain more consistent naming conventions in Shiny apps that use \pkg{bslib}. The documentation on this page may still refer to the original function names. You can replace \code{shiny::checkboxInput()} with \code{input_checkbox()}.
}

\section{Server value}{


\code{TRUE} if checked, \code{FALSE} otherwise.

}

\seealso{
\code{\link[=update_checkbox]{update_checkbox()}} to programmatically update a checkbox.

Other Shiny input aliases: 
\code{\link{input_action_button}()},
\code{\link{input_action_link}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_date}()},
\code{\link{input_date_range}()},
\code{\link{input_numeric}()},
\code{\link{input_password}()},
\code{\link{input_radio_buttons}()},
\code{\link{input_select}()},
\code{\link{input_selectize}()},
\code{\link{input_slider}()},
\code{\link{input_text}()},
\code{\link{input_text_area}()}
}
\concept{Shiny input aliases}
