% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-switch.R
\name{input_switch}
\alias{input_switch}
\alias{update_switch}
\title{Create a switch input}
\usage{
input_switch(id, label, value = FALSE, width = NULL, inline = TRUE)

update_switch(
  id,
  label = NULL,
  value = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{an input id.}

\item{label}{a label for the switch.}

\item{value}{whether or not the switch should be checked by default.}

\item{width}{a valid CSS unit defining the width.}
}
\description{
Create a switch input
}
