% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{input_task_button}
\alias{input_task_button}
\alias{icon_spinner}
\alias{update_task_button}
\title{Button for launching longer-running operations}
\usage{
input_task_button(
  id,
  label,
  icon = NULL,
  busy_label = "Processing...",
  busy_icon = icon_spinner(),
  type = "primary",
  ...
)

icon_spinner()

update_task_button(id, busy = NULL, session = get_current_session())
}
\arguments{
\item{id}{The \code{input} slot that will be used to access the value.}

\item{label}{The label of the button while it is in ready (clickable) state;
usually a string.}

\item{icon}{An optional icon to display next to the label while the button is
in ready state. See \code{\link[fontawesome:fa_i]{fontawesome::fa_i()}}.}

\item{busy_label}{The label of the button while it is busy.}

\item{busy_icon}{The icon to display while the button is busy. By default, a
spinning "chasing arrows" icon is used.}

\item{type}{One of the Bootstrap theme colors (\code{"primary"}, \code{"default"},
\code{"secondary"}, \code{"success"}, \code{"danger"}, \code{"warning"}, \code{"info"}, \code{"light"},
\code{"dark"}), or \code{NULL} to leave off the Bootstrap-specific button CSS classes
altogether.}

\item{...}{Named arguments become attributes to include on the \verb{<button>}
element.}

\item{busy}{If \code{TRUE}, put the button into busy/disabled state. If \code{FALSE},
put the button into ready/enabled state.}

\item{session}{The \code{session} object; using the default is recommended.}
}
\description{
\code{input_task_button} is a button whose value is initially zero, and increments
by one each time it is pressed. It is similar to \code{\link[shiny:actionButton]{shiny::actionButton()}},
except that upon click, it puts itself into a busy state that prevents the
user from clicking it again. \code{update_task_button(id, busy = FALSE)} must be
called from the server function to re-enable the button.
}
\section{Server value}{

An integer of class \code{"shinyActionButtonValue"}. This class differs from
ordinary integers in that a value of 0 is considered "falsy".
This implies two things:
\itemize{
\item Event handlers (e.g., \code{\link[shiny:observeEvent]{shiny::observeEvent()}}, \code{\link[shiny:observeEvent]{shiny::eventReactive()}}) won't execute on initial load.
\item Input validation (e.g., \code{\link[shiny:req]{shiny::req()}}, \code{\link[shiny:validate]{shiny::need()}}) will fail on initial load.
}
}

