% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_columns}
\alias{layout_columns}
\title{Bootstrap grid layout}
\usage{
layout_columns(
  ...,
  widths = NA,
  gap = "1rem",
  fill = TRUE,
  fillable = TRUE,
  height = NULL,
  height_mobile = NULL,
  class = NULL
)
}
\description{
TODO: describe me
}
