% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{layout_sidebar}
\alias{layout_sidebar}
\title{Create a "full bleed" sidebar layout}
\usage{
layout_sidebar(
  ...,
  sidebar = list(),
  sidewidth = 250,
  title = NULL,
  collapsible = TRUE,
  fill = FALSE,
  bg_colors = c("var(--bs-body-bg)", "var(--bs-gray-200)"),
  class = NULL
)
}
\arguments{
\item{fill}{whether or not the \code{main} content area should be considered a fill (i.e., flexbox) container.}

\item{bg_colors}{A character vector of color codes of length 2. The first color is
used for the \code{side} content and the second in is used for the \code{main} content.}

\item{class}{Additional CSS classes for the top-level HTML element.}

\item{side}{A \code{\link[htmltools:builder]{htmltools::tag()}} child to place in the sidebar.}

\item{main}{A \code{\link[htmltools:builder]{htmltools::tag()}} child to place in the main content area.}

\item{side_width}{A valid \link[htmltools:validateCssUnit]{CSS unit} used for the width of the sidebar.}
}
\description{
Create a "full bleed" sidebar layout
}
