% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{layout_sidebar}
\alias{layout_sidebar}
\title{Create a "full bleed" sidebar layout}
\usage{
layout_sidebar(
  side,
  main,
  side_width = "20\%",
  bg_colors = c("var(--bs-body-bg)", "var(--bs-light)"),
  top = "55px"
)
}
\description{
Create a "full bleed" sidebar layout
}
