% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav-items.R, R/navs-legacy.R
\name{nav_item}
\alias{nav_item}
\alias{nav_spacer}
\alias{nav}
\alias{nav_menu}
\alias{nav_content}
\title{Navigation items}
\usage{
nav_item(...)

nav_spacer()

nav(title, ..., value = title, icon = NULL)

nav_menu(title, ..., value = title, icon = NULL, align = c("left", "right"))

nav_content(value, ..., icon = NULL)
}
\arguments{
\item{...}{UI elements (i.e., \link{tags}) to show when the nav is selected ( for
\code{nav_item()}, \code{...} defines the actual nav item, not the selected content).}

\item{title}{A title to display. Can be a character string or UI elements
(i.e., \link{tags}).}

\item{value}{A character string to assign to the nav item. This value may be
supplied to the relevant container's \code{selected} argument in order to show
particular nav item's content immediately on page load. This value is also
useful for programmatically updating the selected content via
\code{\link[=nav_select]{nav_select()}}, \code{\link[=nav_hide]{nav_hide()}}, etc (updating selected tabs this way is often
useful for showing/hiding panels of content via other UI controls like
\code{\link[shiny:radioButtons]{shiny::radioButtons()}} -- in this scenario, consider using \code{\link[=nav_content]{nav_content()}}
with \code{\link[=navs_hidden]{navs_hidden()}}).}

\item{icon}{Optional icon to appear next to the nav item's \code{title}.}

\item{align}{horizontal alignment of the dropdown menu relative to dropdown toggle.}
}
\value{
A nav item that may be passed to a nav container (e.g. \code{\link[=navs_tab]{navs_tab()}}).
}
\description{
Create nav item(s) for use inside nav containers (e.g., \code{\link[=navs_tab]{navs_tab()}},
\code{\link[=navs_bar]{navs_bar()}}, etc).
}
\examples{

navs_tab(
  nav("A", "a"),
  nav("B", "b"),
  nav_menu(
    "More",
    nav("C", "c"),
    "-----",
    nav("D", "d")
  )
)

plot_tag <- function(expr) {
  htmltools::plotTag(
    expr, alt = "R plot", width = 800,
    suppressSize = "x",
    attribs = list(width = "100\%")
  )
}

navs_pill(
  nav(
    "Volcano", plot_tag(image(volcano)),
    icon = if (require("shiny")) icon("r-project")
   ),
  nav(
    "Pressure", plot_tag(plot(pressure)),
    icon = if (require("shiny")) icon("github")
   )
)

}
\seealso{
\code{\link[=navs_tab]{navs_tab()}}, \code{\link[=nav_select]{nav_select()}}.
}
