% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navs-legacy.R
\name{navbar_options}
\alias{navbar_options}
\title{Create a set of navbar options}
\usage{
navbar_options(
  ...,
  position = c("static-top", "fixed-top", "fixed-bottom"),
  bg = NULL,
  type = c("auto", "light", "dark"),
  collapsible = TRUE,
  underline = TRUE
)
}
\arguments{
\item{...}{Additional attributes that will be passed directly to the navbar
container element.}

\item{position}{Determines whether the navbar should be displayed at the top
of the page with normal scrolling behavior (\code{"static-top"}), pinned at
the top (\code{"fixed-top"}), or pinned at the bottom
(\code{"fixed-bottom"}). Note that using \code{"fixed-top"} or
\code{"fixed-bottom"} will cause the navbar to overlay your body content,
unless you add padding, e.g.: \code{tags$style(type="text/css", "body
  {padding-top: 70px;}")}}

\item{bg}{a CSS color to use for the navbar's background color.}

\item{type}{Either \code{"dark"} for a light text color (on a \strong{dark} background)
or \code{"light"} for a dark text color (on a \strong{light} background). If \code{"auto"}
(the default) and \code{bg} is provided, the best contrast to \code{bg} is chosen.}

\item{collapsible}{\code{TRUE} to automatically collapse the navigation
elements into an expandable menu on mobile devices or narrow window widths.}

\item{underline}{Whether or not to add underline styling to page or navbar
links when active or focused.}
}
\value{
Returns a list of navbar options.
}
\description{
A \code{navbar_options()} object captures options specific to the appearance and
behavior of the navbar, independent from the content displayed on the page.
This helper should be used to create the list of options expected by
\code{navbar_options} in \code{\link[=page_navbar]{page_navbar()}} and \code{\link[=navset_bar]{navset_bar()}}.
}
\details{
\subsection{Changelog}{

This function was introduced in \pkg{bslib} v0.9.0, replacing the \code{position},
\code{bg}, \code{inverse}, \code{collapsible} and \code{underline} arguments of \code{\link[=page_navbar]{page_navbar()}}
and \code{\link[=navset_bar]{navset_bar()}}. Those arguments are deprecated with a warning and will be
removed in a future version of \pkg{bslib}. Note that the deprecated
\code{inverse} argument of \code{\link[=page_navbar]{page_navbar()}} and \code{\link[=navset_bar]{navset_bar()}} was replaced with
the \code{type} argument of \code{navbar_options()}.
}
}
\examples{
navbar_options(position = "static-top", bg = "#2e9f7d", underline = FALSE)

}
