% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navs-legacy.R, R/navs.R
\name{navs_tab}
\alias{navs_tab}
\alias{navs_pill}
\alias{navs_pill_list}
\alias{navs_hidden}
\alias{navs_bar}
\alias{navs_tab_card}
\alias{navs_pill_card}
\alias{navs_pill_list_menu}
\title{Create a navigation panel}
\usage{
navs_tab(..., id = NULL, selected = NULL, header = NULL, footer = NULL)

navs_pill(..., id = NULL, selected = NULL, header = NULL, footer = NULL)

navs_pill_list(
  ...,
  id = NULL,
  selected = NULL,
  header = NULL,
  footer = NULL,
  well = TRUE,
  fluid = TRUE,
  widths = c(4, 8)
)

navs_hidden(..., id = NULL, selected = NULL, header = NULL, footer = NULL)

navs_bar(
  ...,
  title = NULL,
  id = NULL,
  selected = NULL,
  position = c("static-top", "fixed-top", "fixed-bottom"),
  header = NULL,
  footer = NULL,
  collapsible = TRUE,
  bg = NULL
)

navs_tab_card(..., id = NULL, selected = NULL, header = NULL, footer = NULL)

navs_pill_card(..., id = NULL, selected = NULL, header = NULL, footer = NULL)

navs_pill_list_menu(
  ...,
  id = NULL,
  selected = NULL,
  header = NULL,
  footer = NULL
)
}
\arguments{
\item{...}{\code{\link[shiny:tabPanel]{tabPanel()}} elements to include in the tabset}

\item{id}{If provided, you can use \verb{input$}\emph{\code{id}} in your
server logic to determine which of the current tabs is active. The value
will correspond to the \code{value} argument that is passed to
\code{\link[shiny:tabPanel]{tabPanel()}}.}

\item{selected}{The \code{value} (or, if none was supplied, the \code{title})
of the tab that should be selected by default. If \code{NULL}, the first
tab will be selected.}

\item{header}{Tag or list of tags to display as a common header above all
tabPanels.}

\item{footer}{Tag or list of tags to display as a common footer below all
tabPanels}

\item{fluid}{\code{TRUE} to use a fluid layout. \code{FALSE} to use a fixed
layout.}

\item{title}{The title to display in the navbar}

\item{position}{This argument is deprecated; it has been discontinued in
Bootstrap 3.}

\item{collapsible}{\code{TRUE} to automatically collapse the navigation
elements into a menu when the width of the browser is less than 940 pixels
(useful for viewing on smaller touchscreen device)}
}
\description{
Render a collection of \code{\link[=nav]{nav()}} items into a panel of contents.
}
