% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page}
\alias{page}
\alias{page_fluid}
\alias{page_fixed}
\alias{page_fill}
\alias{page_navbar}
\title{Create a Bootstrap page}
\usage{
page(..., title = NULL, theme = bs_theme(), lang = NULL)

page_fluid(..., title = NULL, theme = bs_theme(), lang = NULL)

page_fixed(..., title = NULL, theme = bs_theme(), lang = NULL)

page_fill(..., padding = 0, title = NULL, theme = bs_theme(), lang = NULL)

page_navbar(
  ...,
  title = NULL,
  id = NULL,
  selected = NULL,
  position = c("static-top", "fixed-top", "fixed-bottom"),
  header = NULL,
  footer = NULL,
  bg = NULL,
  inverse = "auto",
  collapsible = TRUE,
  fluid = TRUE,
  theme = bs_theme(),
  window_title = NA,
  lang = NULL
)
}
\arguments{
\item{...}{a collection of \code{\link[=nav]{nav()}} items.}

\item{title}{A (left-aligned) title to place in the card header/footer. If
provided, other nav items are automatically right aligned.}

\item{padding}{Whether vertical and/or horizontal padding should be included
around contents. Use \code{NULL} for neither, \code{"x"} for horizontal only,
\code{"y"} for vertical only, or \code{c("x", "y")} for both (the default).}

\item{id}{a character string used for dynamically updating the container (see \code{\link[=nav_select]{nav_select()}}).}

\item{selected}{a character string matching the \code{value} of a particular \code{\link[=nav]{nav()}} item to selected by default.}

\item{header}{UI element(s) (\link{tags}) to display \emph{above} the nav content.}

\item{footer}{UI element(s) (\link{tags}) to display \emph{below} the nav content.}

\item{bg}{a CSS color to use for the navbar's background color.}

\item{inverse}{Either \code{TRUE} for a light text color or \code{FALSE} for a dark
text color. If \code{"auto"} (the default), the best contrast to \code{bg} is chosen.}

\item{window_title}{the browser window title. The default value, \code{NA}, means
to use any character strings that appear in \code{title} (if none are found, the
host URL of the page is displayed by default).}
}
\description{
These functions are small wrappers around shiny's page constructors (i.e., \code{\link[shiny:fluidPage]{shiny::fluidPage()}}, \code{\link[shiny:navbarPage]{shiny::navbarPage()}}, etc) that differ in two ways:
\itemize{
\item The \code{theme} parameter defaults bslib's recommended version of Bootstrap (for new projects).
\item The return value is rendered as an static HTML page when printed interactively at the console.
}
}
\seealso{
\code{\link[shiny:bootstrapPage]{shiny::bootstrapPage()}}

\code{\link[shiny:fluidPage]{shiny::fluidPage()}}

\code{\link[shiny:fixedPage]{shiny::fixedPage()}}

\code{\link[shiny:fillPage]{shiny::fillPage()}}

\code{\link[shiny:navbarPage]{shiny::navbarPage()}}
}
