% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompiled.R
\name{precompiled_css_path}
\alias{precompiled_css_path}
\title{Get the path to a precompiled CSS file}
\usage{
precompiled_css_path(theme = bs_theme())
}
\arguments{
\item{theme}{A \code{\link[=bs_theme]{bs_theme()}} object.}
}
\value{
The path to the precompiled CSS file, if it exists, or \code{NULL}, if
not.
}
\description{
This function is meant for development and debugging purposes. It can be used
to test if a precompiled CSS file exists for a given theme object.
}
\keyword{internal}
