% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{bootstrap_sass_files}
\alias{bootstrap_sass_files}
\alias{bootswatch_sass_files}
\title{Obtain Bootstrap and Bootswatch SASS files}
\usage{
bootstrap_sass_files(files = NULL, version = version_default())

bootswatch_sass_files(theme, files = NULL, version = version_default())
}
\arguments{
\item{version}{the major version.}

\item{theme}{a bootswatch theme name.}

\item{file}{a scss file path.}
}
\description{
Useful if you need to import specific SASS files from
Bootstrap and/or Bootswatch, but hopefully you won't need
this level of control.
}
