% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{sass_file_bootstrap}
\alias{sass_file_bootstrap}
\alias{sass_file_bootswatch}
\title{Obtain Bootstrap and Bootswatch SASS files}
\usage{
sass_file_bootstrap(file = NULL, version = version_default())

sass_file_bootswatch(theme, file = NULL, version = version_default())
}
\arguments{
\item{file}{a scss file path.}

\item{version}{the major version.}

\item{theme}{a bootswatch theme name.}
}
\description{
Useful if you need to import specific SASS files from
Bootstrap and/or Bootswatch, but hopefully you won't need
this level of control. If you're looking to obtain the SASS
behind a theme object, use \code{\link[=bs_theme_sass]{bs_theme_sass()}}.
}
\details{
\code{sass_file_bootstrap()} defaults to the main Bootstrap scss file.
\code{sass_file_bootswatch()} defaults to the variables scss file.
}
\seealso{
\code{\link[=bs_theme_sass]{bs_theme_sass()}}
}
