% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebar}
\alias{sidebar}
\alias{layout_sidebar}
\alias{sidebar_open}
\alias{sidebar_close}
\title{Create various sidebar-based layouts}
\usage{
sidebar(
  ...,
  width = 250,
  collapsible = TRUE,
  id = NULL,
  bg = NULL,
  class = NULL
)

layout_sidebar(
  sidebar = sidebar(),
  ...,
  full_bleed = FALSE,
  fill = FALSE,
  bg = "var(--bs-body-bg)",
  border = !full_bleed,
  border_radius = !full_bleed,
  class = NULL
)

sidebar_open(id, session = get_current_session())

sidebar_close(id, session = get_current_session())
}
\arguments{
\item{...}{A collection of \code{\link[htmltools:builder]{htmltools::tag()}} children to place in the main
content area.}

\item{width}{A valid \link[htmltools:validateCssUnit]{CSS unit} used for the
width of the sidebar.}

\item{collapsible}{Whether or not the sidebar should be collapsible.}

\item{id}{A character string. Required if wanting to re-actively read (or
update) the \code{collapsible} state in a Shiny app.}

\item{bg}{A background color.}

\item{class}{Additional CSS classes for the top-level HTML element.}

\item{sidebar}{A \code{\link[=sidebar]{sidebar()}} object.}

\item{full_bleed}{whether or not to clip the layout container the entire viewport.}

\item{fill}{whether or not the \code{main} content area should be considered a
fill (i.e., flexbox) container.}

\item{border}{whether or not to add a border.}

\item{border_radius}{whether or not to add a border radius.}
}
\description{
Create various sidebar-based layouts
}
\section{Functions}{
\itemize{
\item \code{layout_sidebar()}: A 'low-level' sidebar layout

\item \code{sidebar_open()}: Close a (\code{collapsible}) \code{\link[=sidebar]{sidebar()}}.

\item \code{sidebar_close()}: Close a (\code{collapsible}) \code{\link[=sidebar]{sidebar()}}.

}}
\seealso{
\code{\link[=card_sidebar]{card_sidebar()}}, \code{\link[=container]{container()}}, \code{\link[=page_navbar]{page_navbar()}}
}
