% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebar}
\alias{sidebar}
\alias{layout_sidebar}
\alias{layout_sidebar_full_bleed}
\alias{sidebar_open}
\alias{sidebar_close}
\title{Create various sidebar-based layouts}
\usage{
sidebar(
  ...,
  width = 250,
  collapsible = TRUE,
  id = NULL,
  bg = NULL,
  class = NULL
)

layout_sidebar(
  sidebar = sidebar(),
  ...,
  fill = FALSE,
  bg = NULL,
  border = TRUE,
  border_radius = TRUE,
  height = NULL,
  class = NULL
)

layout_sidebar_full_bleed(
  sidebar = sidebar(),
  ...,
  fill = FALSE,
  bg = NULL,
  inset = NULL,
  class = NULL
)

sidebar_open(id, session = get_current_session())

sidebar_close(id, session = get_current_session())
}
\arguments{
\item{...}{A collection of \code{\link[htmltools:builder]{htmltools::tag()}} children (i.e., UI elements).}

\item{width}{A valid \link[htmltools:validateCssUnit]{CSS unit} used for the
width of the sidebar.}

\item{collapsible}{Whether or not the sidebar should be collapsible.}

\item{id}{A character string. Required if wanting to re-actively read (or
update) the \code{collapsible} state in a Shiny app.}

\item{bg}{A background color.}

\item{class}{Additional CSS classes for the top-level HTML element.}

\item{sidebar}{A \code{\link[=sidebar]{sidebar()}} object.}

\item{fill}{Whether or not the \code{main} content area should be considered a
fill (i.e., flexbox) container.}

\item{border}{Whether or not to add a border.}

\item{border_radius}{Whether or not to add a border radius.}

\item{height}{Any valid \link[htmltools:validateCssUnit]{CSS unit} (e.g.,
\code{height="200px"}).}

\item{inset}{A valid \href{https://developer.mozilla.org/en-US/docs/Web/CSS/inset}{CSS inset} definition.
If not provided, a sensible default to avoid overlap with \code{\link[=page_navbar]{page_navbar()}}
is provided.}

\item{session}{a shiny session object (the default should almost always be
used).}
}
\description{
Create various sidebar-based layouts
}
\section{Functions}{
\itemize{
\item \code{layout_sidebar()}: A 'localized' sidebar layout

\item \code{layout_sidebar_full_bleed()}: A 'full-bleed' sidebar layout

\item \code{sidebar_open()}: Close a (\code{collapsible}) \code{\link[=sidebar]{sidebar()}}.

\item \code{sidebar_close()}: Close a (\code{collapsible}) \code{\link[=sidebar]{sidebar()}}.

}}
\seealso{
\code{\link[=card_sidebar]{card_sidebar()}}, \code{\link[=page_navbar]{page_navbar()}}, \code{\link[=navs_tab_card]{navs_tab_card()}}
}
