% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebar}
\alias{sidebar}
\alias{layout_sidebar}
\alias{sidebar_open}
\alias{sidebar_close}
\title{Sidebar layouts}
\usage{
sidebar(
  ...,
  width = 250,
  position = c("left", "right"),
  open = TRUE,
  id = NULL,
  bg = NULL,
  class = NULL
)

layout_sidebar(
  sidebar = sidebar(),
  ...,
  fillable = FALSE,
  fill = TRUE,
  bg = NULL,
  border = TRUE,
  border_radius = TRUE,
  height = NULL
)

sidebar_open(id, session = get_current_session())

sidebar_close(id, session = get_current_session())
}
\arguments{
\item{...}{A collection of \code{\link[htmltools:builder]{htmltools::tag()}} children (i.e., UI elements).}

\item{width}{A valid \link[htmltools:validateCssUnit]{CSS unit} used for the
width of the sidebar.}

\item{position}{Where the sidebar should appear relative to the main content.}

\item{open}{Whether or not the sidebar should be open on page load. Provide a
value of \code{NA} to prevent sidebar from being collapsible.}

\item{id}{A character string. Required if wanting to re-actively read (or
update) the \code{collapsible} state in a Shiny app.}

\item{bg}{A background color.}

\item{class}{Additional CSS classes for the top-level HTML element.}

\item{sidebar}{A \code{\link[=sidebar]{sidebar()}} object.}

\item{fillable}{Whether or not the \code{main} content area should be considered a
fillable (i.e., flexbox) container.}

\item{fill}{Whether or not the layout container should grow/shrink to fit
a fillable container.}

\item{border}{Whether or not to add a border.}

\item{border_radius}{Whether or not to add a border radius.}

\item{height}{Any valid \link[htmltools:validateCssUnit]{CSS unit} (e.g.,
\code{height="200px"}).}

\item{session}{a shiny session object (the default should almost always be
used).}
}
\description{
Create various sidebar-based layouts. See \href{https://rstudio.github.io/bslib/articles/sidebars.html}{this article} to learn
more.
}
\section{Functions}{
\itemize{
\item \code{layout_sidebar()}: A 'localized' sidebar layout

\item \code{sidebar_open()}: Close a (\code{collapsible}) \code{\link[=sidebar]{sidebar()}}.

\item \code{sidebar_close()}: Close a (\code{collapsible}) \code{\link[=sidebar]{sidebar()}}.

}}
\seealso{
\code{\link[=card_sidebar]{card_sidebar()}}, \code{\link[=page_navbar]{page_navbar()}}, \code{\link[=navs_tab_card]{navs_tab_card()}}
}
