% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebar}
\alias{sidebar}
\alias{layout_sidebar}
\alias{sidebar_open}
\alias{sidebar_close}
\title{Sidebar layouts}
\usage{
sidebar(
  ...,
  width = 250,
  position = c("left", "right"),
  open = TRUE,
  id = NULL,
  bg = NULL,
  class = NULL
)

layout_sidebar(
  sidebar,
  ...,
  fillable = FALSE,
  fill = TRUE,
  bg = NULL,
  border = NULL,
  border_radius = NULL,
  height = NULL
)

sidebar_open(id, session = get_current_session())

sidebar_close(id, session = get_current_session())
}
\arguments{
\item{...}{Unnamed arguments can be any valid child of an \link[htmltools:builder]{htmltools tag} and named arguments become HTML attributes on
returned UI element. In the case of \code{layout_sidebar()}, these arguments are
passed to the main content tag (not the sidebar+main content container).}

\item{width}{A valid \link[htmltools:validateCssUnit]{CSS unit} used for the
width of the sidebar.}

\item{position}{Where the sidebar should appear relative to the main content.}

\item{open}{Whether or not the sidebar should be open on page load. Provide a
value of \code{NA} to prevent sidebar from being collapsible.}

\item{id}{A character string. Required if wanting to re-actively read (or
update) the \code{collapsible} state in a Shiny app.}

\item{bg}{A background color. If provided, an accessible contrasting color is
provided for the foreground color (consider using a utility \code{class} to
customize the foreground color).}

\item{class}{Additional CSS classes for the top-level HTML element.}

\item{sidebar}{A \code{\link[=sidebar]{sidebar()}} object.}

\item{fillable}{Whether or not the \code{main} content area should be considered a
fillable (i.e., flexbox) container.}

\item{fill}{Whether or not the layout container should grow/shrink to fit
a fillable container.}

\item{border}{Whether or not to add a border.}

\item{border_radius}{Whether or not to add a border radius.}

\item{height}{Any valid \link[htmltools:validateCssUnit]{CSS unit} (e.g.,
\code{height="200px"}).}

\item{session}{a shiny session object (the default should almost always be
used).}
}
\description{
Create a collapsing sidebar layout by providing a \code{sidebar()}
object to the \code{sidebar} argument of:
\itemize{
\item \code{layout_sidebar()}
\itemize{
\item Creates a sidebar layout component which can be dropped inside any
\code{\link[=page]{page()}} or \code{\link[=card]{card()}} context.
}
\item \code{\link[=page_navbar]{page_navbar()}}, \code{\link[=navs_tab_card]{navs_tab_card()}}, and \code{\link[=navs_pill_card]{navs_pill_card()}}
\itemize{
\item Creates a multi page/tab UI with a singular \code{sidebar()} (which is
shown on every page/tab).
}
}

See \href{https://rstudio.github.io/bslib/articles/sidebars.html}{this article}
to learn more.
}
\section{Functions}{
\itemize{
\item \code{sidebar_open()}: Open a \code{sidebar()} (during an active Shiny user session).

\item \code{sidebar_close()}: Close a \code{sidebar()} (during an active Shiny user session).

}}
