% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootswatch.R
\name{theme_bootswatch}
\alias{theme_bootswatch}
\title{Obtain a theme's Bootswatch theme name}
\usage{
theme_bootswatch(theme)
}
\arguments{
\item{theme}{a \code{\link[=bs_theme]{bs_theme()}} object.}
}
\value{
the Bootswatch theme named used (if any) in the \code{theme}.
}
\description{
Obtain a theme's Bootswatch theme name
}
