% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-theme-preset-bootswatch.R
\name{theme_bootswatch}
\alias{theme_bootswatch}
\title{Obtain a theme's Bootswatch theme name}
\usage{
theme_bootswatch(theme)
}
\arguments{
\item{theme}{a \code{\link[=bs_theme]{bs_theme()}} object.}
}
\value{
Returns the Bootswatch theme named used (if any) in the \code{theme}.
}
\description{
Obtain a theme's Bootswatch theme name
}
\seealso{
Other Bootstrap theme utility functions: 
\code{\link{bootswatch_themes}()},
\code{\link{bs_get_variables}()},
\code{\link{builtin_themes}()},
\code{\link{theme_version}()},
\code{\link{versions}()}
}
\concept{Bootstrap theme utility functions}
