% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootswatch.R
\name{theme_version}
\alias{theme_version}
\title{Obtain a theme's Bootstrap version}
\usage{
theme_version(theme = bs_theme_get())
}
\arguments{
\item{theme}{a bs theme object, see \code{\link[=bs_theme_set]{bs_theme_set()}}.}
}
\description{
Obtain a theme's Bootstrap version
}
