% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-theme-preset-bootswatch.R
\name{theme_version}
\alias{theme_version}
\title{Obtain a theme's Bootstrap version}
\usage{
theme_version(theme)
}
\arguments{
\item{theme}{A \code{\link[=bs_theme]{bs_theme()}} object.}
}
\value{
Returns the major version of Bootstrap used in the \code{theme}.
}
\description{
Obtain a theme's Bootstrap version
}
\seealso{
Other Bootstrap theme utility functions: 
\code{\link{bootswatch_themes}()},
\code{\link{bs_get_variables}()},
\code{\link{builtin_themes}()},
\code{\link{theme_bootswatch}()},
\code{\link{versions}()}
}
\concept{Bootstrap theme utility functions}
