% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{bs_theme}
\alias{bs_theme}
\alias{bs_theme_set}
\alias{bs_theme_get}
\alias{bs_theme_clear}
\alias{bs_theme_merge}
\alias{bs_theme_sass}
\title{Create a Bootstrap theme}
\usage{
bs_theme(
  pre = "",
  post = "",
  bootswatch = NULL,
  version = c("4-3", "4", "3"),
  deps = NULL
)

bs_theme_set(
  pre = "",
  post = "",
  bootswatch = NULL,
  version = c("4-3", "4", "3"),
  deps = NULL
)

bs_theme_get()

bs_theme_clear()

bs_theme_merge(...)

bs_theme_sass(theme = NULL, pre_only = TRUE)
}
\arguments{
\item{pre}{Any \code{\link[sass:as_sass]{sass::as_sass()}} input to place before Bootstrap's SASS code.
For \code{bs_theme_set()}, this may also be a \code{bs_theme()} object (in that case,
no other arguments may be used).}

\item{post}{Any \code{\link[sass:as_sass]{sass::as_sass()}} input to place after Bootstrap's SASS code.}

\item{bootswatch}{The name of a bootswatch theme.
See \code{\link[=bootswatch_themes]{bootswatch_themes()}} for a list of possible names.}

\item{version}{The major version of Bootstrap to use. A value of
\code{'4-3'} means Bootstrap 4, but with additional CSS/JS to support
BS3 style markup in BS4. Other supported versions include 3 and 4.}

\item{deps}{An HTML dependency (or a list of them).}
}
\description{
Use \code{\link[=bs_theme]{bs_theme()}} (or \code{\link[=bs_theme_set]{bs_theme_set()}}) to create a customized
Bootstrap SASS theme which \code{\link[=bs_sass]{bs_sass()}} (or \code{\link[=bs_sass_partial]{bs_sass_partial()}}) can consume.
Either pass a \code{\link[=bs_theme]{bs_theme()}} object directly to \code{\link[=bs_sass]{bs_sass()}}'s \code{theme} argument
or use \code{\link[=bs_theme_set]{bs_theme_set()}} to set a theme globally (\code{\link[=bs_sass]{bs_sass()}}'s \code{theme} argument
defaults to \code{bs_theme_get()}, which get the current global).
}
\details{
Developers who need their own sass compilation entry point may use the
\code{bs_theme_sass()}
}
\examples{

# Override Bootstrap SASS variables (e.g., set $primary to red)
red_primary <- bs_theme(list("primary" = "red !default;"))
foo_color <- ".foo { color: rgba($primary, 0.3) }"
bs_sass_partial(foo_color)
bs_sass_partial(foo_color, theme = red_primary)

# To create new variables that reference Bootstrap variables,
# those
my_primary <- list("my-primary" = "$primary !default;")
bs_sass_partial(
  ".bar { color: $my-primary }",
  theme = bs_theme(post = my_primary)
)

# For complex theming projects, it may be useful to combine
# individually packaged themes via bs_theme_merge() which
# combines several theme into one by layering them
# together like an onion -- the pre sass for latter themes
# (e.g., blue_primary) is placed _before_ pre sass for former
# themes (e.g., red_primary) and post sass for latter themes
# is placed _after_ the post sass for former themes

blue_primary <- bs_theme(list("primary" = "blue !default;"))
primary <- bs_theme_merge(red_primary, blue_primary)
# The blue theme wins out
bs_sass_partial(foo_color, theme = primary)

}
\references{
\url{https://getbootstrap.com/docs/4.4/getting-started/theming/}
}
