% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbar.R
\name{toolbar}
\alias{toolbar}
\alias{toolbar_divider}
\title{Toolbar component}
\usage{
toolbar(..., align = c("right", "left"), gap = NULL)

toolbar_divider(..., width = NULL, gap = NULL)
}
\arguments{
\item{...}{UI elements for the toolbar.}

\item{align}{Determines if toolbar should be aligned to the \code{"right"} or
\code{"left"}.}

\item{gap}{A CSS length unit defining the spacing around the divider.
Defaults to \code{"1rem"} for sensible fixed spacing.}

\item{width}{A CSS length unit specifying the width of the divider line.
Defaults to \code{"2px"} for a sensible dividing line. Pass \verb{0px} for no
divider line.}
}
\value{
Returns a toolbar element.
}
\description{
A toolbar which can contain buttons, inputs, and other UI elements in a small
form suitable for inclusion in card headers, footers, and other small places.

bslib includes a complete example of toolbars and the many ways they can be used:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{shiny::runExample("toolbar", package = "bslib")
}\if{html}{\out{</div>}}

\code{toolbar_divider()} creates a visual divider line with customizable width
and spacing between toolbar elements.
}
\section{Functions}{
\itemize{
\item \code{toolbar()}: Create a toolbar.

\item \code{toolbar_divider()}: Add a divider to a toolbar.

}}
\examples{
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}
toolbar(
  align = "right",
  toolbar_input_button(id = "see", icon = icon("eye"), label = "View"),
  toolbar_input_button(id = "save", icon = icon("save"), label = "Save"),
  toolbar_input_button(id = "edit", icon = icon("pencil"), label = "Edit"),
  toolbar_divider(),
  toolbar_input_select(
    id = "select",
    label = "Choose option",
    choices = c("Option 1", "Option 2", "Option 3"),
    selected = "Option 2"
  )
)
\dontshow{\}) # examplesIf}
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}
toolbar(
  toolbar_input_button(id = "left1", label = "Left"),
  toolbar_divider(),
  toolbar_input_button(id = "right1", label = "Right")
)

toolbar(
  toolbar_input_button(id = "a", label = "A"),
  toolbar_divider(width = "5px", gap = "20px"),
  toolbar_input_button(id = "b", label = "B")
)
\dontshow{\}) # examplesIf}
}
\seealso{
\url{toolbar_input_button}, \url{toolbar_input_select},
\url{toolbar_divider} for other Toolbar components.

\code{\link[=toolbar]{toolbar()}} to create a toolbar, \code{\link[=toolbar_input_button]{toolbar_input_button()}} to create a toolbar button, \code{\link[=toolbar_input_select]{toolbar_input_select()}} to create a toolbar select input.

Other toolbar components: 
\code{\link{toolbar_input_button}()},
\code{\link{toolbar_input_select}()}

Other toolbar components: 
\code{\link{toolbar_input_button}()},
\code{\link{toolbar_input_select}()}
}
\concept{toolbar components}
