% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbar.R
\name{toolbar}
\alias{toolbar}
\title{Add a toolbar to a UI element}
\usage{
toolbar(..., align = c("right", "left"), size = c("sm", "md", "lg"))
}
\arguments{
\item{...}{UI elements for the toolbar.}

\item{align}{Determines if toolbar should be aligned to the right or left.
Must be one of "right" or "left".}

\item{size}{The size of the toolbar. Must be one of "sm", "md", or "lg".}
}
\value{
Returns a toolbar.
}
\description{
A toolbar which can contain buttons, inputs, and other UI elements
in a small form suitable for inclusion in card headers,
footers, and other small places.
}
