% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbar.R
\name{toolbar}
\alias{toolbar}
\title{Toolbar component}
\usage{
toolbar(..., align = c("right", "left"), gap = NULL, width = "auto")
}
\arguments{
\item{...}{UI elements for the toolbar.}

\item{align}{Determines if toolbar should be aligned to the \code{"right"} or
\code{"left"}.}

\item{gap}{A CSS length unit defining the gap (i.e., spacing) between
elements in the toolbar. Defaults to \code{0} (no gap).}

\item{width}{CSS width of the toolbar. Default is \code{auto}.}
}
\value{
Returns a toolbar element.
}
\description{
A toolbar which can contain buttons, inputs, and other UI elements in a small
form suitable for inclusion in card headers, footers, and other small places.

bslib includes a complete example of toolbars and the many ways they can be used:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{shiny::runExample("toolbar", package = "bslib")
}\if{html}{\out{</div>}}
}
\section{Toolbars in labels}{


You can use toolbars in the labels of other Shiny inputs to add a composite
input with additional buttons or controls. The following example uses a
\code{toolbar()} in the label of a \code{\link[shiny:numericInput]{shiny::numericInput()}} to add increment and
decrement buttons next to the label:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{shiny::numericInput(
  "quantity",
  label = toolbar(
    "Quantity",
    toolbar_spacer(), # push buttons to the right
    toolbar_input_button("decrement", "Less", shiny::icon("minus")),
    toolbar_input_button("increment", "More", shiny::icon("plus"))
  ),
  value = 1,
  min = 0,
  max = 100
)
}\if{html}{\out{</div>}}
}

\examples{
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}
toolbar(
  align = "right",
  toolbar_input_button(id = "see", icon = icon("eye"), label = "View"),
  toolbar_input_button(id = "save", icon = icon("save"), label = "Save"),
  toolbar_input_button(id = "edit", icon = icon("pencil"), label = "Edit"),
  toolbar_divider(),
  toolbar_input_select(
    id = "select",
    label = "Choose option",
    choices = c("Option 1", "Option 2", "Option 3"),
    selected = "Option 2"
  )
)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=card_header]{card_header()}} for using toolbars in card headers/footers

Other toolbar components: 
\code{\link{toolbar_divider}()},
\code{\link{toolbar_input_button}()},
\code{\link{toolbar_input_select}()}
}
\concept{toolbar components}
