% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbar.R
\name{toolbar}
\alias{toolbar}
\title{Toolbar component}
\usage{
toolbar(..., align = c("right", "left"))
}
\arguments{
\item{...}{UI elements for the toolbar.}

\item{align}{Determines if toolbar should be aligned to the \code{"right"} or
\code{"left"}.}
}
\value{
Returns a toolbar element.
}
\description{
A toolbar which can contain buttons, inputs, and other UI elements in a small
form suitable for inclusion in card headers, footers, and other small places.
}
