% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbar.R
\name{toolbar}
\alias{toolbar}
\title{Add a toolbar to a UI element}
\usage{
toolbar(..., align = c("right", "left"), size = c("sm", "md", "lg"))
}
\arguments{
\item{...}{UI elements for the toolbar.}

\item{align}{A character string.}

\item{size}{STUFF HERE}
}
\value{
Returns a toolbar container.
}
\description{
Display additional information when focusing (or hovering over) a UI element.
}
