% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbar.R
\name{toolbar}
\alias{toolbar}
\title{Toolbar component}
\usage{
toolbar(..., align = c("right", "left"), gap = NULL)
}
\arguments{
\item{...}{UI elements for the toolbar.}

\item{align}{Determines if toolbar should be aligned to the \code{"right"} or
\code{"left"}.}

\item{gap}{A CSS length unit defining the gap (i.e., spacing) between
elements in the toolbar. Defaults to \code{0} (no gap).}
}
\value{
Returns a toolbar element.
}
\description{
A toolbar which can contain buttons, inputs, and other UI elements in a small
form suitable for inclusion in card headers, footers, and other small places.
}
\examples{
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}
toolbar(
  align = "right",
  toolbar_input_button(id = "see", icon = icon("eye"), label = "View"),
  toolbar_input_button(id = "save", icon = icon("save"), label = "Save"),
  toolbar_input_button(id = "edit", icon = icon("pencil"), label = "Edit")
)
\dontshow{\}) # examplesIf}
}
\seealso{
Other Toolbar components: 
\code{\link{toolbar_input_button}()}
}
\concept{Toolbar components}
