% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbar.R
\name{toolbar_divider}
\alias{toolbar_divider}
\alias{toolbar_spacer}
\title{Toolbar: Add a divider or spacer to a toolbar}
\usage{
toolbar_divider(..., width = NULL, gap = NULL)

toolbar_spacer()
}
\arguments{
\item{...}{Ignored, reserved for future use and to require named arguments
in \code{toolbar_divier()}.}

\item{width}{A CSS length unit specifying the width of the divider line.
Defaults to \code{"2px"} for a sensible dividing line. Pass \verb{0px} for no
divider line.}

\item{gap}{A CSS length unit defining the spacing around the divider.
Defaults to \code{"1rem"} for sensible fixed spacing.}
}
\description{
\code{toolbar_divider()} creates a visual divider line with customizable and fixed
width and spacing between toolbar elements. \code{toolbar_spacer()} creates empty
space that expands to push adjacent toolbar elements apart as much as
possible.
}
\section{Functions}{
\itemize{
\item \code{toolbar_divider()}: Create a dividing line and fixed space between
toolbar elements.

\item \code{toolbar_spacer()}: Create empty, expanding space between toolbar
elements. Note that for the spacer to push elements effectively, the parent
toolbar needs \verb{width: 100\%} (which is automatically applied when the toolbar
is a direct child of a label element). If the spacer doesn't appear to work,
you may need to set \code{width = "100\%"} explicitly on the \code{\link[=toolbar]{toolbar()}}.

}}
\examples{
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}
toolbar(
  toolbar_input_button(id = "left1", label = "Left"),
  toolbar_divider(),
  toolbar_input_button(id = "right1", label = "Right")
)

toolbar(
  toolbar_input_button(id = "a", label = "A"),
  toolbar_divider(width = "5px", gap = "20px"),
  toolbar_input_button(id = "b", label = "B")
)

toolbar(
  toolbar_input_button(id = "previous", label = "Previous"),
  toolbar_spacer(),
  toolbar_input_button(id = "next", label = "Next")
)
\dontshow{\}) # examplesIf}
}
\seealso{
Other toolbar components: 
\code{\link{toolbar}()},
\code{\link{toolbar_input_button}()},
\code{\link{toolbar_input_select}()}
}
\concept{toolbar components}
