% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbar.R
\name{toolbar_input_button}
\alias{toolbar_input_button}
\title{Add toolbar button input}
\usage{
toolbar_input_button(
  id,
  label,
  icon = NULL,
  show_label = is.null(icon),
  tooltip = !show_label,
  ...,
  disabled = FALSE,
  border = FALSE
)
}
\arguments{
\item{id}{The input ID.}

\item{label}{The input label. By default, \code{label} is not shown but is used by
\code{tooltip}. Set \code{show_label = TRUE} to show the label (see \code{tooltip} for
details on how this affects the tooltip behavior).}

\item{icon}{An icon. If provided without \code{show_label = TRUE}, only the icon
will be visible.}

\item{show_label}{Whether to show the label text. If \code{FALSE} (the default),
only the icon is shown (if provided). If \code{TRUE}, the label text is shown
alongside the icon.}

\item{tooltip}{Tooltip text to display when hovering over the input. Can be:
\itemize{
\item \code{TRUE} (default when \code{show_label = FALSE}) - shows a tooltip with the
\code{label} text
\item \code{FALSE} (default when \code{show_label = TRUE}) - no tooltip
\item A character string - shows a tooltip with custom text Defaults to
\code{!show_label}.
}}

\item{...}{Additional attributes to pass to the button.}

\item{disabled}{If \code{TRUE}, the button will not be clickable. Use
\code{\link[shiny:updateActionButton]{shiny::updateActionButton()}} to dynamically enable/disable the button.}

\item{border}{Whether to show a border around the button.}
}
\value{
Returns a button suitable for use in a toolbar.
}
\description{
A button designed to fit well in small places such as in a \code{\link[=toolbar]{toolbar()}}.
}
\examples{
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}
toolbar(
  align = "right",
  toolbar_input_button(id = "see", icon = icon("eye"), label = "See"),
  toolbar_input_button(id = "save", label = "Save"),
  toolbar_input_button(id = "edit", icon = icon("pencil"), label = "Edit", show_label = TRUE)
)
\dontshow{\}) # examplesIf}
}
\seealso{
Other Toolbar components: 
\code{\link{toolbar}()},
\code{\link{toolbar_divider}()},
\code{\link{toolbar_input_select}()}
}
\concept{Toolbar components}
