% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbar.R
\name{toolbar_input_select}
\alias{toolbar_input_select}
\title{Toolbar Input Select}
\usage{
toolbar_input_select(id, choices, selected = NULL, width = NULL)
}
\arguments{
\item{id}{The input ID.}

\item{choices}{List of values to select from. If elements of the list are
named, then that name - rather than the value - is displayed to the user.
It's also possible to group related inputs by providing a named list whose
elements are (either named or unnamed) lists, vectors, or factors. In this
case, the outermost names will be used as the group labels (leveraging the
\verb{<optgroup>} HTML tag) for the elements in the respective sublist.}

\item{selected}{The initially selected value. If not specified then defaults
to the first value.}

\item{width}{The width of the input, e.g. '400px', or '100\%'; see
\code{\link[htmltools:validateCssUnit]{htmltools::validateCssUnit()}}.}
}
\value{
Returns a select input control suitable for use in a toolbar.
}
\description{
Create a select list input control that can be used to choose a single item
from a list of values, suitable for use within a \code{\link[=toolbar]{toolbar()}}. This is a
wrapper around \code{\link[shiny:selectInput]{shiny::selectInput()}} with \code{selectize = FALSE} and
appropriate styling for toolbar usage.
}
\examples{
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}
toolbar(
  align = "right",
  toolbar_input_select(
    id = "select",
    choices = c("Option 1", "Option 2", "Option 3"),
    selected = "Option 2"
  )
)
\dontshow{\}) # examplesIf}
}
\concept{Toolbar components}
