% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbar.R
\name{toolbar_input_switch}
\alias{toolbar_input_switch}
\title{Add toolbar switch input}
\usage{
toolbar_input_switch(id, label = NULL, value = FALSE)
}
\arguments{
\item{id}{The input ID.}

\item{label}{The label to display next to the switch. If \code{NULL}, no label
is displayed.}

\item{value}{The initial state of the switch. Default is \code{FALSE}.}
}
\value{
Returns a switch input suitable for use in a toolbar.
}
\description{
A switch input designed to fit well in small places such as in a \code{\link[=toolbar]{toolbar()}}.
#'   toolbar_input_switch(id = "toggle", label = "Enable", value = TRUE),
toolbar_input_switch(id = "switch", value = FALSE)
)
}
\seealso{
Other Toolbar components: 
\code{\link{toolbar}()},
\code{\link{toolbar_divider}()},
\code{\link{toolbar_input_button}()},
\code{\link{toolbar_input_select}()}
}
\concept{Toolbar components}
