% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip}
\alias{tooltip}
\alias{tooltip_toggle}
\alias{tooltip_update}
\title{Add a tooltip to a UI element}
\usage{
tooltip(
  trigger,
  ...,
  id = NULL,
  placement = c("auto", "top", "right", "bottom", "left"),
  options = list()
)

tooltip_toggle(id, show = NULL, session = get_current_session())

tooltip_update(id, ..., session = get_current_session())
}
\arguments{
\item{trigger}{A UI element (i.e., \link[htmltools:builder]{htmltools tag}) to serve as the
tooltips trigger. It's good practice for this element to be a keyboard-focusable
and interactive element (e.g., \code{actionButton()}, \code{actionLink()}, etc) so that
the tooltip is accessible to keyboard and assistive technology users.}

\item{...}{UI elements for the tooltip. Character strings are automatically
\code{\link[=htmlEscape]{htmlEscape()}}d unless marked as \code{\link[=HTML]{HTML()}}.}

\item{id}{a character string that matches an existing tooltip id.}

\item{placement}{The placement of the tooltip relative to its trigger.}

\item{options}{A list of additional \href{https://getbootstrap.com/docs/5.3/components/tooltips/#options}{Bootstrap options}.}

\item{show}{Whether to show (\code{TRUE}) or hide (\code{FALSE}) the tooltip. The
default (\code{NULL}) will show if currently hidden and hide if currently shown.}

\item{session}{A Shiny session object (the default should almost always be
used).}
}
\description{
Tooltips are useful for showing additional information when focusing (or
hovering over) a UI element.
}
\details{
If \code{x} yields multiple HTML elements, the last element is
used as the tooltip's target. If the target should contain multiple elements,
then wrap those elements in a \code{\link[=span]{span()}} or \code{\link[=div]{div()}}.
}
\section{Functions}{
\itemize{
\item \code{tooltip()}: Add a tooltip to a UI element

\item \code{tooltip_toggle()}: Programmatically show/hide a tooltip.

\item \code{tooltip_update()}: Update the contents of a tooltip.

}}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

tooltip(
  shiny::actionButton("btn", "A button"),
  "A message"
)

card(
  card_header(
    tooltip(
      span("Card title ", bsicons::bs_icon("question-circle-fill")),
      "Additional info",
      placement = "right"
    )
  ),
  "Card body content..."
)
\dontshow{\}) # examplesIf}
}
\references{
\url{https://getbootstrap.com/docs/5.3/components/tooltips/}
}
