% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip}
\alias{tooltip}
\alias{toggle_tooltip}
\alias{update_tooltip}
\title{Add a tooltip to a UI element}
\usage{
tooltip(
  trigger,
  ...,
  id = NULL,
  placement = c("auto", "top", "right", "bottom", "left"),
  options = list()
)

toggle_tooltip(id, show = NULL, session = get_current_session())

update_tooltip(id, ..., session = get_current_session())
}
\arguments{
\item{trigger}{A UI element (i.e., \link[htmltools:builder]{htmltools tag}) to serve
as the tooltips trigger. It's good practice for this element to be a
keyboard-focusable and interactive element (e.g., \code{actionButton()},
\code{actionLink()}, etc) so that the tooltip is accessible to keyboard and
assistive technology users.}

\item{...}{UI elements for the tooltip. Character strings are \link[=htmlEscape]{automatically escaped} unless marked as \code{\link[=HTML]{HTML()}}.}

\item{id}{a character string that matches an existing tooltip id.}

\item{placement}{The placement of the tooltip relative to its trigger.}

\item{options}{A list of additional \href{https://getbootstrap.com/docs/5.3/components/tooltips/#options}{Bootstrap options}.}

\item{show}{Whether to show (\code{TRUE}) or hide (\code{FALSE}) the tooltip. The
default (\code{NULL}) will show if currently hidden and hide if currently shown.
Note that a tooltip will not be shown if the trigger is not visible (e.g.,
it's hidden behind a tab).}

\item{session}{A Shiny session object (the default should almost always be
used).}
}
\description{
Display additional information when focusing (or hovering over) a UI element.
}
\details{
If \code{trigger} yields multiple HTML elements (e.g., a \code{tagList()} or
complex \code{{htmlwidgets}} object), the last HTML element is used as the
trigger. If the \code{trigger} should contain all of those elements, wrap the
object in a \code{\link[=div]{div()}} or \code{\link[=span]{span()}}.
}
\section{Functions}{
\itemize{
\item \code{tooltip()}: Add a tooltip to a UI element

\item \code{toggle_tooltip()}: Programmatically show/hide a tooltip.

\item \code{update_tooltip()}: Update the contents of a tooltip.

}}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

tooltip(
  shiny::actionButton("btn", "A button"),
  "A message"
)

card(
  card_header(
    tooltip(
      span("Card title ", bsicons::bs_icon("question-circle-fill")),
      "Additional info",
      placement = "right"
    )
  ),
  "Card body content..."
)
\dontshow{\}) # examplesIf}
}
\references{
\url{https://getbootstrap.com/docs/5.3/components/tooltips/}
}
\seealso{
\code{\link[=popover]{popover()}}
}
