% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value-box.R
\name{value_box}
\alias{value_box}
\alias{value_box_theme}
\alias{showcase_left_center}
\alias{showcase_top_right}
\alias{showcase_bottom}
\title{Value box}
\usage{
value_box(
  title,
  value,
  ...,
  showcase = NULL,
  showcase_layout = c("left center", "top right", "bottom"),
  full_screen = FALSE,
  theme = NULL,
  height = NULL,
  max_height = NULL,
  fill = TRUE,
  class = NULL,
  theme_color = lifecycle::deprecated()
)

value_box_theme(name = NULL, bg = NULL, fg = NULL)

showcase_left_center(
  width = 0.3,
  max_height = "100px",
  max_height_full_screen = 0.67,
  width_full_screen = "1fr"
)

showcase_top_right(
  width = 0.4,
  max_height = "75px",
  max_height_full_screen = 0.67,
  width_full_screen = "1fr"
)

showcase_bottom(
  width = "100\%",
  width_full_screen = NULL,
  height = "auto",
  height_full_screen = "2fr",
  max_height = "100px",
  max_height_full_screen = NULL
)
}
\arguments{
\item{title, value}{A string, number, or \code{\link[htmltools:builder]{htmltools::tag()}} child to display as
the title or value of the value box. The \code{title} appears above the \code{value}.}

\item{...}{Unnamed arguments may be any \code{\link[htmltools:builder]{htmltools::tag()}} children to
display below \code{value}. Named arguments become attributes on the containing
element.}

\item{showcase}{A \code{\link[htmltools:builder]{htmltools::tag()}} child to showcase (e.g., a
\code{\link[bsicons:bs_icon]{bsicons::bs_icon()}}, a \code{plotly::plotlyOutput()}, etc).}

\item{showcase_layout}{One of \code{"left center"} (default), \code{"top right"} or
\code{"bottom"}. Alternatively, you can customize the showcase layout options
with the \code{\link[=showcase_left_center]{showcase_left_center()}}, \code{\link[=showcase_top_right]{showcase_top_right()}}, or
\code{\link[=showcase_bottom]{showcase_bottom()}} functions. Use the options functions when you want to
control the height or width of the showcase area.}

\item{full_screen}{If \code{TRUE}, an icon will appear when hovering over the card
body. Clicking the icon expands the card to fit viewport size.}

\item{theme}{The name of a theme for the value box, or a theme constructed
with \code{value_box_theme()}. The theme names provide a convenient way to use
your app's Bootstrap theme colors as the foreground or background colors of
the value box. See below for more details on the provided themes. For more
control, you can create your own theme with \code{value_box_theme()} where you
can pass foreground and background colors directly.

There are many prepared theme options available to you:
\enumerate{
\item \strong{Named themes.} Choose from the names of Bootstrap theme colors (from
\verb{$theme-colors}, e.g. \code{primary}, \code{secondary}, \code{success}, \code{danger}, etc)
or named main colors (from \verb{$colors}, e.g. \code{purple}, \code{orange}, \code{blue},
etc.).
\item \strong{Background colors.} If the theme or color name is provided without
any prefix, the color will be used for the background of the value box.
You can also explicitly prefix the theme or color name with \verb{bg-} to
indicate that it should apply to the value box background. An
appropriate text color is chosen automatically.
\item \strong{Foreground colors.} If the theme or color name is prefixed with
\verb{text-}, the color will be be applied to the foreground of the value box
and the background color remains the default card background color.
\item \strong{Gradient backgrounds.} You can also choose a vibrant gradient
background for the value box by using a theme name of the form
\code{bg-gradient-{from}-{to}} where \code{{from}} and \code{{to}} are named main
colors, e.g. \code{bg-gradient-purple-orange}.
\item \strong{Further customization.} For further customization, you can always set
the \code{theme} and add additional classes using the \code{class} argument. If
you want to create your own custom themes, for best results be sure to
include the \verb{text-} or \verb{bg-} prefix in your theme's class.
}}

\item{max_height, max_height_full_screen}{A proportion (i.e., a number between
0 and 1) or any valid \link[htmltools:validateCssUnit]{CSS unit} defining the
showcase max_height.}

\item{fill}{Whether to allow the value box to grow/shrink to fit a fillable
container with an opinionated height (e.g., \code{page_fillable()}).}

\item{class}{Utility classes for customizing the appearance of the summary
card. Use \verb{bg-*} and \verb{text-*} classes (e.g, \code{"bg-danger"} and
\code{"text-light"}) to customize the background/foreground colors.}

\item{theme_color}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{theme} instead.}

\item{name}{The name of the theme, e.g. \code{"primary"}, \code{"danger"}, \code{"purple"}.}

\item{bg, fg}{The background and foreground colors for the theme. If only \code{bg}
is provided, then the foreground color is automatically chosen from
\verb{$black} or \verb{$white} to provide the best contrast with the background
color.}

\item{width, width_full_screen, height, height_full_screen}{one of the
following:
\itemize{
\item A proportion (i.e., a number between 0 and 1) of available width or
height to allocate to the showcase.
\item A valid \link[htmltools:validateCssUnit]{CSS unit} defining the width or
height of the showcase column, or a valid value accepted by the
\code{grid-template-columns} (width) or \code{grid-template-rows} (height) CSS
property to define the width or height of the showcase column or row.
Accepted values in the second category are \code{"auto"}, \code{"min-content"},
\code{"max-content"}, a fractional unit (e.g. \verb{2fr}), or a \code{minmax()} function
(e.g., \verb{minmax(100px, 1fr)}).
}}
}
\description{
An opinionated (\code{\link[=card]{card()}}-powered) box, designed for displaying a \code{value} and
\code{title}. Optionally, a \code{showcase} can provide for context for what the
\code{value} represents (for example, it could hold a \code{\link[bsicons:bs_icon]{bsicons::bs_icon()}}, or
even a \code{\link[shiny:plotOutput]{shiny::plotOutput()}}).
}
\examples{
library(htmltools)

if (interactive()) {
  value_box(
    "KPI Title",
    h1(HTML("$1 <i>Billion</i> Dollars")),
    span(
      bsicons::bs_icon("arrow-up"),
      " 30\% VS PREVIOUS 30 DAYS"
    ),
    showcase = bsicons::bs_icon("piggy-bank"),
    class = "bg-success"
  )
}

}
\seealso{
\code{\link[=card]{card()}}
}
