% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version-default.R
\name{versions}
\alias{versions}
\alias{version_default}
\title{Available Bootstrap versions}
\usage{
versions()

version_default()
}
\value{
A list of the Bootstrap versions available.
}
\description{
Available Bootstrap versions
}
